fix(SFT-2077): add csrf auto-fetch for all forms
