+++ packages/plugin/src/Resources/js/scripts/front-end/plugin/freeform.js
@@ -1,2 +1,2 @@
 /*! For license information please see freeform.js.LICENSE.txt */
-!function(){"use strict";var e={d:function(t,r){for(var n in r)e.o(r,n)&&!e.o(t,n)&&Object.defineProperty(t,n,{enumerable:!0,get:r[n]})},o:function(e,t){return Object.prototype.hasOwnProperty.call(e,t)}};e.d({},{A:function(){return me}});var t,r={ready:"freeform-ready",reset:"freeform-on-reset",submit:"freeform-on-submit",removeMessages:"freeform-remove-messages",fieldRemoveMessages:"freeform-remove-field-messages",renderSuccess:"freeform-render-success",renderFieldErrors:"freeform-render-field-errors",renderFormErrors:"freeform-render-form-errors",ajaxBeforeSuccess:"freeform-before-ajax-success",ajaxSuccess:"freeform-ajax-success",ajaxError:"freeform-ajax-error",ajaxBeforeSubmit:"freeform-ajax-before-submit",ajaxAfterSubmit:"freeform-ajax-after-submit",afterFailedSubmit:"freeform-after-failed-submit",handleActions:"freeform-handle-actions"},n={applied:"freeform-rules-applied"},o={onAddRow:"freeform-field-table-on-add-row",afterRowAdded:"freeform-field-table-after-row-added",onRemoveRow:"freeform-field-table-on-remove-row",afterRemoveRow:"freeform-field-table-after-remove-row"},a={renderPreview:"freeform-field-dnd-on-render-preview",renderPreviewRemoveButton:"freeform-field-dnd-on-render-preview-remove-button",renderErrorContainer:"freeform-field-dnd-render-error-container",showGlobalMessage:"freeform-field-dnd-show-global-message",appendErrors:"freeform-field-dnd-append-errors",clearErrors:"freeform-field-dnd-clear-errors",onChange:"freeform-field-dnd-on-change",onUploadProgress:"freeform-field-dnd-on-upload-progress"},i={saveFormhandleToken:"freeform-save-form-handle-token"};!function(e){e.Reload="reload",e.RedirectReturnUrl="redirect-return-url",e.LoadSuccessTemplate="load-success-template"}(t||(t={}));var s,l=function(e){var t=this;this.reload=function(){for(var e=t.form.querySelectorAll("*[name=".concat(me._BACK_BUTTON_NAME,"]")),r=0;r<e.length;r++)e[r].addEventListener("click",(function(){var e=document.createElement("input");e.type="hidden",e.name=me._BACK_BUTTON_NAME,e.value="",t.form.appendChild(e)}))},this.form=e.form,this.reload()},c=function(){return c=Object.assign||function(e){for(var t,r=1,n=arguments.length;r<n;r++)for(var o in t=arguments[r])Object.prototype.hasOwnProperty.call(t,o)&&(e[o]=t[o]);return e},c.apply(this,arguments)},d=function(e){var t=this;if(this.loadedLocales={},this.scriptAdded=!1,this.reload=function(){t.freeform.has("data-scripts-datepicker")&&t.freeform.form.querySelectorAll("*[data-datepicker][data-datepicker-enabled]").forEach((function(e){var r=e.getAttribute("data-datepicker-locale"),n={disableMobile:!0,allowInput:!0,dateFormat:e.getAttribute("data-datepicker-format"),enableTime:null!==e.getAttribute("data-datepicker-enabletime"),noCalendar:null===e.getAttribute("data-datepicker-enabledate"),time_24hr:null!==e.getAttribute("data-datepicker-clock_24h"),minDate:e.getAttribute("data-datepicker-min-date"),maxDate:e.getAttribute("data-datepicker-max-date"),minuteIncrement:1,hourIncrement:1,static:null!==e.getAttribute("data-datepicker-static")},o=t.freeform._dispatchEvent("flatpickr-before-init",{detail:n,options:n}),a=c(c({},o.detail),o.options),i=flatpickr(e,a);if(e.setAttribute("autocomplete","off"),t.freeform._dispatchEvent("flatpickr-ready",{detail:i,flatpickr:i}),t.loadedLocales[r])t.loadedLocales[r].addEventListener("load",(function(){i.set("locale",r),t.loadedLocales[r].dataset.loaded="true"})),"true"===t.loadedLocales[r].dataset.loaded&&i.set("locale",r);else{var s=document.createElement("script");s.src="//cdnjs.cloudflare.com/ajax/libs/flatpickr/4.6.13/l10n/".concat(r,".js"),s.async=!1,s.defer=!1,s.dataset.loaded="",s.addEventListener("load",(function(){i.set("locale",r),s.dataset.loaded="true"})),document.body.appendChild(s),t.loadedLocales[r]=s}}))},this.freeform=e,this.freeform.has("data-scripts-datepicker")&&!this.scriptAdded){var r=document.createElement("script");r.src="//cdnjs.cloudflare.com/ajax/libs/flatpickr/4.6.13/flatpickr.min.js",r.async=!1,r.defer=!1,r.addEventListener("load",(function(){t.reload()})),document.body.appendChild(r);var n=document.createElement("link");n.rel="stylesheet",n.href="//cdnjs.cloudflare.com/ajax/libs/flatpickr/4.6.13/flatpickr.min.css",document.body.appendChild(n),this.scriptAdded=!0}},u=function(e,t,r){var n=t||{},o=n.bubbles,a=void 0!==o&&o,i=n.cancelable,s=void 0===i||i,l=function(e,t){var r={};for(var n in e)Object.prototype.hasOwnProperty.call(e,n)&&t.indexOf(n)<0&&(r[n]=e[n]);if(null!=e&&"function"==typeof Object.getOwnPropertySymbols){var o=0;for(n=Object.getOwnPropertySymbols(e);o<n.length;o++)t.indexOf(n[o])<0&&Object.prototype.propertyIsEnumerable.call(e,n[o])&&(r[n[o]]=e[n[o]])}return r}(n,["bubbles","cancelable"]),c=f(e,a,s);return Object.assign(c,l),r&&(r instanceof HTMLElement?r.dispatchEvent(c):Array.from(r).forEach((function(e){return e.dispatchEvent(c)}))),c},f=function(e,t,r){return void 0===t&&(t=!0),void 0===r&&(r=!0),new Event(e,{bubbles:t,cancelable:r})},p=(s=function(e,t){return s=Object.setPrototypeOf||{__proto__:[]}instanceof Array&&function(e,t){e.__proto__=t}||function(e,t){for(var r in t)Object.prototype.hasOwnProperty.call(t,r)&&(e[r]=t[r])},s(e,t)},function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Class extends value "+String(t)+" is not a constructor or null");function r(){this.constructor=e}s(e,t),e.prototype=null===t?Object.create(t):(r.prototype=t.prototype,new r)}),m=function(){return m=Object.assign||function(e){for(var t,r=1,n=arguments.length;r<n;r++)for(var o in t=arguments[r])Object.prototype.hasOwnProperty.call(t,o)&&(e[o]=t[o]);return e},m.apply(this,arguments)},h=function(e,t,r,n){return new(r||(r=Promise))((function(o,a){function i(e){try{l(n.next(e))}catch(e){a(e)}}function s(e){try{l(n.throw(e))}catch(e){a(e)}}function l(e){var t;e.done?o(e.value):(t=e.value,t instanceof r?t:new r((function(e){e(t)}))).then(i,s)}l((n=n.apply(e,t||[])).next())}))},v=function(e,t){var r,n,o,a={label:0,sent:function(){if(1&o[0])throw o[1];return o[1]},trys:[],ops:[]},i=Object.create(("function"==typeof Iterator?Iterator:Object).prototype);return i.next=s(0),i.throw=s(1),i.return=s(2),"function"==typeof Symbol&&(i[Symbol.iterator]=function(){return this}),i;function s(s){return function(l){return function(s){if(r)throw new TypeError("Generator is already executing.");for(;i&&(i=0,s[0]&&(a=0)),a;)try{if(r=1,n&&(o=2&s[0]?n.return:s[0]?n.throw||((o=n.return)&&o.call(n),0):n.next)&&!(o=o.call(n,s[1])).done)return o;switch(n=0,o&&(s=[2&s[0],o.value]),s[0]){case 0:case 1:o=s;break;case 4:return a.label++,{value:s[1],done:!1};case 5:a.label++,n=s[1],s=[0];continue;case 7:s=a.ops.pop(),a.trys.pop();continue;default:if(!((o=(o=a.trys).length>0&&o[o.length-1])||6!==s[0]&&2!==s[0])){a=0;continue}if(3===s[0]&&(!o||s[1]>o[0]&&s[1]<o[3])){a.label=s[1];break}if(6===s[0]&&a.label<o[1]){a.label=o[1],o=s;break}if(o&&a.label<o[2]){a.label=o[2],a.ops.push(s);break}o[2]&&a.ops.pop(),a.trys.pop();continue}s=t.call(e,a)}catch(e){s=[6,e],n=0}finally{r=o=0}if(5&s[0])throw s[1];return{value:s[0]?s[1]:void 0,done:!0}}([s,l])}}},b=function(e,t,r){if(r||2===arguments.length)for(var n,o=0,a=t.length;o<a;o++)!n&&o in t||(n||(n=Array.prototype.slice.call(t,0,o)),n[o]=t[o]);return e.concat(n||Array.prototype.slice.call(t))},g=function(e,t,r,n,o){var a=new XMLHttpRequest;return a.open(e,t),a.setRequestHeader("Cache-Control","no-cache"),a.setRequestHeader("X-Requested-With","XMLHttpRequest"),a.setRequestHeader("HTTP_X_REQUESTED_WITH","XMLHttpRequest"),y(a,null==o?void 0:o.headers),a.onload=function(){var e=a.response;try{e=JSON.parse(a.response)}catch(e){}var t=a.status;t<200||t>=300?n(new S("Request failed with status ".concat(a.statusText),a,e)):r({status:a.status,data:e})},a.onerror=function(){n(new Error("Network error"))},a.onabort=function(){n(new Error("Request aborted"))},o.onUploadProgress&&(a.upload.onprogress=function(e){o.onUploadProgress(e)}),o.cancelToken&&o.cancelToken._setCancelFn((function(){a.abort()})),a},y=function(e,t){t&&Object.entries(t).forEach((function(t){var r=t[0],n=t[1];e.setRequestHeader(r,String(n))}))},E=function(e,t){return new Promise((function(r,n){var o=g((null==t?void 0:t.method)||"GET",e,r,n,t),a=null==t?void 0:t.data;a instanceof FormData?o.send(a):(o.setRequestHeader("Content-Type","application/json"),o.send(JSON.stringify(a)))}))};E.get=function(e){for(var t=[],r=1;r<arguments.length;r++)t[r-1]=arguments[r];return h(void 0,b([e],t,!0),void 0,(function(e,t){return void 0===t&&(t={}),v(this,(function(r){return[2,new Promise((function(r,n){var o=g("GET",e,r,n,t);o.open("GET",e),o.send()}))]}))}))},E.post=function(e,t){for(var r=[],n=2;n<arguments.length;n++)r[n-2]=arguments[n];return h(void 0,b([e,t],r,!0),void 0,(function(e,t,r){return void 0===r&&(r={}),v(this,(function(n){return[2,new Promise((function(n,o){var a=g("POST",e,n,o,r);t instanceof FormData?a.send(t):(a.setRequestHeader("Content-Type","application/json"),a.send(JSON.stringify(t)))}))]}))}))};var w=function(){function e(){}return e.prototype.cancel=function(){this.cancelFn&&(this.cancelFn(),this.cancelFn=null)},e.prototype._setCancelFn=function(e){this.cancelFn=e},e}(),S=function(e){function t(r,n,o){var a=e.call(this,r)||this;return a.response=m(m({},n),{data:o}),a.status=n.status,Object.setPrototypeOf(a,t.prototype),a}return p(t,e),t}(Error);const A="bits",_="bytes",k="jedec",L={symbol:{iec:{bits:["bit","Kibit","Mibit","Gibit","Tibit","Pibit","Eibit","Zibit","Yibit"],bytes:["B","KiB","MiB","GiB","TiB","PiB","EiB","ZiB","YiB"]},jedec:{bits:["bit","Kbit","Mbit","Gbit","Tbit","Pbit","Ebit","Zbit","Ybit"],bytes:["B","KB","MB","GB","TB","PB","EB","ZB","YB"]}},fullform:{iec:["","kibi","mebi","gibi","tebi","pebi","exbi","zebi","yobi"],jedec:["","kilo","mega","giga","tera","peta","exa","zetta","yotta"]}};var C,x=function(e,t,r){if(r||2===arguments.length)for(var n,o=0,a=t.length;o<a;o++)!n&&o in t||(n||(n=Array.prototype.slice.call(t,0,o)),n[o]=t[o]);return e.concat(n||Array.prototype.slice.call(t))},T="freeform-file-dnd",q=function(e){for(var t=[],r=1;r<arguments.length;r++)t[r-1]=arguments[r];e.classList.add(t.join("__"))},F=function(e){for(var t=[],r=1;r<arguments.length;r++)t[r-1]=arguments[r];e.classList.remove(t.join("__"))},j=function(e){for(var t=[],r=1;r<arguments.length;r++)t[r-1]=arguments[r];q.apply(void 0,x([e,T],t,!1))},P=function(e){return"string"==typeof e&&(e=e.split(" ")),e.map((function(e){return".".concat(e)})).join("")},O=function(e,t){var r=function(e){var t=e.name,r=e.extension,n=e.size,o=document.createElement("div");o.setAttribute("data-file-preview",""),j(o,"preview-zone","file-preview");var a=document.createElement("span");a.setAttribute("data-extension-label",""),a.innerText=r.toUpperCase(),j(a,"preview-zone","file-preview","thumbnail","extension-label");var i=document.createElement("div");i.setAttribute("data-thumbnail",""),i.appendChild(a),j(i,"preview-zone","file-preview","thumbnail");var s,l,c,d,u=document.createElement("span");u.setAttribute("data-filename",""),u.innerText=(void 0===(l=14)&&(l=50),void 0===c&&(c="…"),void 0===d&&(d=3),(s=t).length>l?s.substr(0,l-d)+c:s),u.title=t,j(u,"preview-zone","file-preview","filename");var f=document.createElement("span");return f.setAttribute("data-filesize",""),f.innerText=n,j(f,"preview-zone","file-preview","filesize"),o.appendChild(i),o.appendChild(u),o.appendChild(f),[o,i,u,f]}(e),n=r[0],o=r[1];e.url&&(o.style.backgroundImage="url(".concat(e.url,")"));var i=function(e){var t=document.createElement("a");return t.setAttribute("data-remove-button",""),t.setAttribute("title","Remove file?"),t.setAttribute("aria-label","Remove file?"),t.innerHTML='<svg style="height: 14px; width: 14px;" aria-hidden="true" focusable="false" data-prefix="fas" data-icon="times" class="svg-inline--fa fa-times fa-w-11" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 352 512"><path fill="currentColor" d="M242.72 256l100.07-100.07c12.28-12.28 12.28-32.19 0-44.48l-22.24-22.24c-12.28-12.28-32.19-12.28-44.48 0L176 189.28 75.93 89.21c-12.28-12.28-32.19-12.28-44.48 0L9.21 111.45c-12.28 12.28-12.28 32.19 0 44.48L109.28 256 9.21 356.07c-12.28 12.28-12.28 32.19 0 44.48l22.24 22.24c12.28 12.28 32.2 12.28 44.48 0L176 322.72l100.07 100.07c12.28 12.28 32.2 12.28 44.48 0l22.24-22.24c12.28-12.28 12.28-32.19 0-44.48L242.72 256z"></path></svg>',j(t,"preview-zone","file-preview","thumbnail","remove-button"),e._dispatchEvent(a.renderPreviewRemoveButton,{button:t}).button}(t),s=function(e){var t=document.createElement("div");return t.innerText="!",t.setAttribute("data-errors",""),j(t,"preview-zone","file-preview","thumbnail","errors"),e._dispatchEvent(a.renderErrorContainer,{container:t},t).container}(t),l=function(){var e=document.createElement("div");return e.setAttribute("data-progress",""),j(e,"preview-zone","file-preview","thumbnail","progress"),e}();o.appendChild(i),o.appendChild(s),o.appendChild(l);var c=t._dispatchEvent(a.renderPreview,{metadata:e,container:n});return j(n,"preview-zone","file-preview","animation-enter"),setTimeout((function(){!function(e){for(var t=[],r=1;r<arguments.length;r++)t[r-1]=arguments[r];F.apply(void 0,x([e,T],t,!1))}(n,"preview-zone","file-preview","animation-enter")}),10),c.container.addEventListener("click",(function(e){return e.stopPropagation()})),c.container},R=function(e,t){var r=t.id,n=document.createElement("input");return n.type="hidden",n.value=r,n.name="".concat(e,"[]"),n};!function(e){e.FieldError="field-error"}(C||(C={}));var B=["png","jpg","jpeg","gif"],M=function(e){return B.includes(e.toLowerCase())},N=function(e,t,r,n,o){var i=new w,s=function(){i.cancel()},l=e.name.match(/.(\w+)$/i),c=e.name,d=function(e,{bits:t=!1,pad:r=!1,base:n=-1,round:o=2,locale:a="",localeOptions:i={},separator:s="",spacer:l=" ",symbols:c={},standard:d="",output:u="string",fullform:f=!1,fullforms:p=[],exponent:m=-1,roundingMethod:h="round",precision:v=0}={}){let b=m,g=Number(e),y=[],E=0,w="";"si"===d?(n=10,d=k):"iec"===d||d===k?n=2:2===n?d="iec":(n=10,d=k);const S=10===n?1e3:1024,C=!0===f,x=g<0,T=Math[h];if("bigint"!=typeof e&&isNaN(e))throw new TypeError("Invalid number");if("function"!=typeof T)throw new TypeError("Invalid rounding method");if(x&&(g=-g),(-1===b||isNaN(b))&&(b=Math.floor(Math.log(g)/Math.log(S)),b<0&&(b=0)),b>8&&(v>0&&(v+=8-b),b=8),"exponent"===u)return b;if(0===g)y[0]=0,w=y[1]=L.symbol[d][t?A:_][b];else{E=g/(2===n?Math.pow(2,10*b):Math.pow(1e3,b)),t&&(E*=8,E>=S&&b<8&&(E/=S,b++));const e=Math.pow(10,b>0?o:0);y[0]=T(E*e)/e,y[0]===S&&b<8&&-1===m&&(y[0]=1,b++),w=y[1]=10===n&&1===b?t?"kbit":"kB":L.symbol[d][t?A:_][b]}if(x&&(y[0]=-y[0]),v>0&&(y[0]=y[0].toPrecision(v)),y[1]=c[y[1]]||y[1],!0===a?y[0]=y[0].toLocaleString():a.length>0?y[0]=y[0].toLocaleString(a,i):s.length>0&&(y[0]=y[0].toString().replace(".",s)),r&&o>0){const e=y[0].toString(),t=s||(e.match(/(\D)/g)||[]).pop()||".",r=e.toString().split(t),n=r[1]||"",a=n.length,i=o-a;y[0]=`${r[0]}${t}${n.padEnd(a+i,"0")}`}return C&&(y[1]=p[b]?p[b]:L.fullform[d][b]+(t?"bit":"byte")+(1===y[0]?"":"s")),"array"===u?y:"object"===u?{value:y[0],symbol:y[1],exponent:b,unit:w}:y.join(l)}(e.size,{round:1}),f=null!==l?l[1].toLowerCase():"n/a",p=O({name:c,extension:f,size:d},o),m=p.querySelector("[data-thumbnail]"),h=p.querySelector("[data-remove-button]");if(M(f)){var v=new FileReader;v.readAsDataURL(e),v.onloadend=function(){m.setAttribute("data-has-preview",""),m.style.backgroundImage="url(".concat(v.result.toString(),")")}}n.appendChild(p),h.addEventListener("click",s),U(r);var b=new FormData(o.form);b.delete("action"),b.append("handle",t),b.append(t,e);var g=r.getAttribute("data-base-url");return E.post("".concat(g,"/files/upload"),b,{cancelToken:i,onUploadProgress:function(e){var t=e.total,n=e.loaded,o=Math.ceil(n/(t/100));u(a.onUploadProgress,{total:t,loaded:n,percent:o},r),p.style.setProperty("--progress","".concat(o,"%")),o>=98&&h.removeEventListener("click",s)}}).then((function(e){var a=new FormData(o.form);a.delete("action"),a.append("handle",t),a.append("id",e.data.id),h.removeEventListener("click",s),h.addEventListener("click",(function(){confirm("Are you sure?")&&E.post("".concat(g,"/files/delete"),a).then((function(){n.removeChild(p),U(r)})).catch((function(e){alert(e.message)}))})),p.appendChild(R(t,e.data)),p.setAttribute("data-completed","")})).catch((function(e){var t,i,l;if("Request aborted"===e.message)return n.removeChild(p),void U(r);if(h.removeEventListener("click",s),h.addEventListener("click",(function(){n.removeChild(p),U(r)})),(null===(i=null===(t=null==e?void 0:e.response)||void 0===t?void 0:t.data)||void 0===i?void 0:i.type)===C.FieldError){var c=(null===(l=null==e?void 0:e.response)||void 0===l?void 0:l.data).messages;!function(e,t,r,n){if(!n._dispatchEvent(a.clearErrors,{},e).defaultPrevented){var o=t.querySelector("[data-errors]");t.setAttribute("data-has-errors",""),o.setAttribute("aria-label",r.join("; ")),o.setAttribute("title",r.join(". "))}}(r,p,c,o)}else console.warn(e)}))},U=function(e){u(a.onChange,{container:e},e)},H=function(e,t,r){void 0===r&&(r=4e3);var n=document.createElement("li");n.setAttribute("data-error",""),n.innerText=t,j(n,"messages","message"),j(n,"messages","message","error");var o=u(a.showGlobalMessage,{messageItem:n},e);I(e,o.messageItem,r)},I=function(e,t,r){var n=e.querySelector("[data-messages]");n&&(n.appendChild(t),setTimeout((function(){t.setAttribute("data-animate-fade-out","")}),r),setTimeout((function(){n.removeChild(t)}),r+300))},D=function(e){var t=this;this.currentFileUploads=0,this.isFormLocked=!1,this.reload=function(){var e=t.freeform.form;t.freeform.form.querySelectorAll("[data-freeform-file-upload]").forEach((function(n){n.style.setProperty("--accent",n.dataset.accent),n.addEventListener("dragenter",t.handleDrag(n)),n.addEventListener("dragleave",t.handleDragLeave(n)),n.addEventListener("dragover",t.handleDrag(n)),n.addEventListener("drop",t.handleDrop(n)),n.addEventListener("click",t.handleClick(n)),n.addEventListener(a.onChange,t.handleChanges),function(e,t){if(parseInt(e.dataset.fileCount||"0")){var r=e.querySelector("[data-preview-zone]"),n=e.dataset.freeformFileUpload,o=new FormData(t.form);o.delete("action"),o.append("handle",n);var a=e.getAttribute("data-base-url");E.post("".concat(a,"/files"),o,{headers:{"Freeform-Preflight":!0}}).then((function(o){o.data.forEach((function(o){var i=O(o,t);M(o.extension)&&i.querySelector("[data-thumbnail]").setAttribute("data-has-preview","");var s=new FormData(t.form);s.delete("action"),s.append("handle",n),s.append("id",o.id),i.querySelector("[data-remove-button]").addEventListener("click",(function(){confirm("Are you sure?")&&E.post("".concat(a,"/files/delete"),s).then((function(){r.removeChild(i),U(e)})).catch((function(e){alert(e.message)}))})),i.appendChild(R(n,o)),i.setAttribute("data-completed",""),r.appendChild(i)})),U(e)})).catch(console.error)}}(n,t.freeform),e.addEventListener(r.reset,t.handleReset(n)),n.querySelector("input[type=file]").addEventListener("change",t.handleManualUpload(n))}))},this.handleChanges=function(e){var t=e.container;t.querySelector("[data-preview-zone]").querySelectorAll("[data-file-preview]").length>0?t.setAttribute("data-contains-files",""):t.removeAttribute("data-contains-files")},this.attachDragState=function(e){e instanceof HTMLElement&&(e.dataset.dragging="")},this.detachDragState=function(e){e instanceof HTMLElement&&delete e.dataset.dragging},this.handleDrag=function(e){return function(r){r.preventDefault(),r.stopPropagation(),t.attachDragState(e)}},this.handleDragLeave=function(e){return function(r){r.preventDefault(),r.stopPropagation(),t.detachDragState(e)}},this.handleDrop=function(e){return function(r){r.preventDefault(),r.stopPropagation(),t.detachDragState(e);var n=r.dataTransfer.files;t.initFileUpload(n,e)}},this.handleClick=function(e){return function(){var t=e.querySelector('input[type="file"]');if(!t)throw new Error("File upload corrupted");t.click()}},this.handleManualUpload=function(e){return function(r){var n=r.target,o=n.files;t.initFileUpload(o,e),n.value=null}},this.initFileUpload=function(e,r){for(var n=r.dataset,o=n.freeformFileUpload,a=n.maxFiles,i=n.maxSize,s=r.dataset,l=s.messageSize,c=s.messageFiles,d=r.querySelector("[data-preview-zone]"),u=r.querySelectorAll("[data-file-preview]:not([data-has-errors])").length,f=0;f<e.length;f++){if(u>=parseInt(a)){H(r,c);break}var p=e.item(f);p.size>parseInt(i)?H(r,l):(t.currentFileUploads++,N(p,o,r,d,t.freeform).finally((function(){t.currentFileUploads--,t.handleUploadLockdown()})),u++,t.handleUploadLockdown())}},this.handleReset=function(e){return function(){e.querySelectorAll("[data-file-preview]").forEach((function(e){return e.parentNode.removeChild(e)})),u(a.onChange,{container:e},e)}},this.handleUploadLockdown=function(){t.currentFileUploads>0?t.isFormLocked||(t.isFormLocked=!0,t.freeform.lockSubmit("file-upload")):(t.freeform.unlockSubmit("file-upload"),t.isFormLocked=!1)},this.freeform=e,this.reload()},z=function(e){var t=this;if(this.scriptAdded=!1,this.reload=function(){t.freeform.has("data-scripts-js-mask")&&t.freeform.form.querySelectorAll("*[data-masked-input]").forEach((function(e){var t=e.getAttribute("data-pattern");t&&new IMask(e,{mask:t})}))},this.freeform=e,this.freeform.has("data-scripts-js-mask")&&!this.scriptAdded){var r=document.createElement("script");r.src="https://cdnjs.cloudflare.com/ajax/libs/imask/6.0.7/imask.min.js",r.async=!1,r.defer=!1,r.addEventListener("load",(function(){t.reload()})),document.body.appendChild(r),this.scriptAdded=!0}},$=function(e){var t=this;this.reload=function(){t.freeform.form.querySelectorAll('[data-field-type="rating"]').forEach((function(e){var t=e.dataset,r=t.colorIdle,n=t.colorHover,o=t.colorSelected;e.style.setProperty("--ff-rating-color-idle",r),e.style.setProperty("--ff-rating-color-hover",n),e.style.setProperty("--ff-rating-color-selected",o)}))},this.freeform=e,this.reload()},G=function(e){var t=this;if(this.scriptAdded=!1,this.reload=function(){t.freeform.has("data-scripts-signature")&&t.freeform.form.querySelectorAll("canvas[data-signature-field]").forEach((function(e){var t=e.dataset,r=t.borderColor,n=t.backgroundColor,o=t.penColor,a=t.dotSize;e.style.borderWidth="1px",e.style.borderStyle="solid",e.style.borderColor=r;var i=e.previousSibling,s=e.parentNode.querySelector("[data-signature-clear]"),l=i.value,c=new SignaturePad(e,{onEnd:function(){i.value=c.toDataURL()},backgroundColor:n,penColor:o,dotSize:a,maxWidth:a,throttle:5});if(s&&s.addEventListener("click",(function(){c.clear(),i.value=""})),l){var d=new Image;c.clear(),d.src=l,d.onload=function(){e.getContext("2d").drawImage(d,0,0,e.width,e.height)}}}))},this.freeform=e,this.freeform.has("data-scripts-signature")&&!this.scriptAdded){var r=document.createElement("script");r.src="//cdn.jsdelivr.net/npm/signature_pad@2.3.2/dist/signature_pad.min.js",r.async=!1,r.defer=!1,r.addEventListener("load",(function(){t.reload()})),document.body.appendChild(r),this.scriptAdded=!0}},W=/([^[]+)\[(\d+)\](\[\d+\])$/g,X=/^(labeled-.*)-(\d+)-(\d+)-(\d+)$/g,J=function(e,t){var r=e.getAttribute("data-max-rows"),n=e.querySelectorAll("tbody > tr").length;r&&t&&(t.style.display=n>=parseInt(r,10)?"none":"block")},K=function(e){return"string"==typeof e&&(e=e.split(" ")),e},Y=function(e,t){K(t).map((function(t){return e.classList.add(t)}))},Z=function(e,t){K(t).map((function(t){return e.classList.remove(t)}))};function Q(e){var t;void 0!==e&&(e instanceof HTMLElement&&(null===(t=e.parentElement)||void 0===t||t.removeChild(e)),(e instanceof HTMLCollection||e instanceof NodeList)&&Array.from(e).forEach((function(e){var t;null===(t=e.parentElement)||void 0===t||t.removeChild(e)})))}var V=!1,ee=function(e){return function(t){var r=t.target;if(1!==r.closest("tbody").querySelectorAll("tr").length){var n=r.closest("table"),a=r.closest("tr");e._dispatchEvent(o.onRemoveRow,{table:n,row:a}),Q(a),e._dispatchEvent(o.afterRemoveRow,{table:n})}}},te=function(e){var t=e.form.querySelectorAll("[data-freeform-table]"),r=e.freeform.options.errorClassField;t.forEach((function(e){var t=e.querySelectorAll("thead th[data-column-required]"),n=Array.from(t).map((function(e){return e.cellIndex}));e.querySelectorAll("tbody tr").forEach((function(e){var t=e.querySelectorAll("td");n.forEach((function(e){var n=t[e].querySelector("input, textarea, select"),o=Boolean(n.value);["radio","checkbox"].includes((null==n?void 0:n.type)||"")&&(o=n.checked),n&&o?Z(n,r):"radio"===(null==n?void 0:n.type)?n.parentElement.parentElement.querySelectorAll("input").forEach((function(e){Y(e,r)})):Y(n,r)}))}))}))},re=function(e){var t=this;this.reload=function(){var e;(function(e){var t;e.form.querySelectorAll("[data-freeform-table]").forEach((function(r){t=r.parentNode.querySelector("[data-freeform-table-add-row]"),J(r,t),t&&(e.form.addEventListener(o.afterRemoveRow,(function(){J(r,t)})),t.addEventListener("click",(function(){var n=r.querySelector("tbody > tr:last-child");if(n){for(var a=n.cloneNode(!0),i=a.querySelectorAll("textarea, input, select"),s=function(){for(var e=r.querySelectorAll("textarea, input, select"),t=0,n=0;n<e.length;n++){var o=e[n],a=W.exec(o.name);if(W.lastIndex=0,a){var i=parseInt(a[2]);t=Math.max(t,i)}}return++t}(),l=0;l<i.length;l++){var c=i[l],d=c.dataset.defaultValue||"";c.name=c.name.replace(W,"$1[".concat(s,"]$3")),c.id&&X.test(c.id)&&(c.id=c.id.replace(X,"$1-".concat(s,"-$3-$4")),c.nextSibling.htmlFor=c.id),"INPUT"===c.tagName&&"radio"===c.type?c.checked=c.value===d:"INPUT"===c.tagName&&"checkbox"===c.type?c.checked="1"===d:c.value=d}e._dispatchEvent(o.onAddRow,{table:r,row:a}),r.querySelector("tbody").appendChild(a),e._dispatchEvent(o.afterRowAdded,{table:r,row:a}),J(r,t)}})))}))})(t.freeform),(e=t.freeform).form.querySelectorAll("[data-freeform-table]").forEach((function(t){for(var r=t.querySelectorAll("[data-freeform-table-remove-row]"),n=0;n<r.length;n++)r[n].addEventListener("click",ee(e))})),V||(V=!0,e.form.addEventListener(o.afterRowAdded,(function(t){var r=t.row.querySelector("[data-freeform-table-remove-row]");r&&r.addEventListener("click",ee(e))}))),function(e){e.form.addEventListener(r.renderFieldErrors,te)}(t.freeform)},this.freeform=e,this.reload()},ne=function(){return ne=Object.assign||function(e){for(var t,r=1,n=arguments.length;r<n;r++)for(var o in t=arguments[r])Object.prototype.hasOwnProperty.call(t,o)&&(e[o]=t[o]);return e},ne.apply(this,arguments)},oe=function(e){if(this.reload=function(){},window.dataLayer=window.dataLayer||[],this.freeform=e,this.form=e.form,this.freeform.has("data-gtm")){var t=this.form.dataset.gtmEventName||"form-submission",n=this.form.dataset.handle;this.form.addEventListener(r.ajaxSuccess,(function(r){var o=r.response,a=e._dispatchEvent("freeform-gtm-data-layer-push",{payload:{},response:o}),i=ne({event:t,form:n,submission:{id:o.submissionId,token:o.submissionToken}},a.payload);window.dataLayer.push(i)}))}},ae=function(e,t){if(e===t)return!0;if(null==e||null==t)return e===t;if(Number.isNaN(e)&&Number.isNaN(t))return!0;if(typeof e!=typeof t)return!1;if(Array.isArray(e)&&Array.isArray(t))return e.length===t.length&&e.every((function(e,r){return ae(e,t[r])}));if("object"==typeof e&&"object"==typeof t){var r=Object.keys(e),n=Object.keys(t);return r.length===n.length&&r.every((function(r){return t.hasOwnProperty(r)&&ae(e[r],t[r])}))}return!1},ie=function(e,t,r){if(r||2===arguments.length)for(var n,o=0,a=t.length;o<a;o++)!n&&o in t||(n||(n=Array.prototype.slice.call(t,0,o)),n[o]=t[o]);return e.concat(n||Array.prototype.slice.call(t))},se=function(e){return function(t){return t.conditions.some((function(t){var r=e.name,n=t.field;return n===r||"".concat(n,"[]")===r}))}},le=function(e){var t=this;this.reload=function(){var e=t.form.querySelector("[data-rules-json]");if(e){var r=JSON.parse(e.dataset.rulesJson),n=r.rules,o=r.values;t.values=o,0===n.fields.length&&0===n.buttons.length||(Array.from(t.form.elements).forEach((function(e){var r=n.fields.filter(se(e)),o=n.buttons.filter(se(e)),a=ie(ie([],r,!0),o,!0);if(0!==a.length){var i;switch(e.tagName){case"TEXTAREA":case"INPUT":var s=e;if("hidden"===s.type)return;switch(s.type){case"radio":case"checkbox":i="change";break;case"number":i=["keyup","change"];break;default:i="keyup"}break;case"SELECT":i="change"}if(i){var l=Array.isArray(i)?i:[i];a.forEach((function(r){l.forEach((function(o){return e.addEventListener(o,function(e){return function(){t.applyRule(e),n.fields.filter((function(t){return t!==e})).forEach((function(e){return t.applyRule(e)}))}}(r))}))}))}}})),n.fields.forEach((function(e){return t.applyRule(e)})),n.buttons.forEach((function(e){return t.applyRule(e)})))}},this.applyRule=function(e){var r="field"in e?'[data-field-container="'.concat(e.field,'"]'):'[data-button-container="'.concat(e.button,'"]'),o=t.form.querySelector(r);if(o){var a=e.display,i=e.combinator,s=e.conditions,l="and"===i?s.every(t.verifyCondition):s.some(t.verifyCondition);return"show"===a?(o.style.display=l?"":"none",l?delete o.dataset.hidden:o.dataset.hidden=""):(o.style.display=l?"none":"",l?o.dataset.hidden="":delete o.dataset.hidden),u(n.applied,{rule:e},o),!0}},this.verifyCondition=function(e){var r=null,n=document.querySelector('[data-field-container="'.concat(e.field,'"]'));if(n){var o=t.form[e.field]||t.form["".concat(e.field,"[]")],a="checkbox"===n.getAttribute("data-field-type"),i="checkboxes"===n.getAttribute("data-field-type");r=void 0!==n.dataset.hidden?null:a?o[1].checked?"1":"":i?Array.from(o).filter((function(e){return e.checked})).map((function(e){return e.value})):o instanceof HTMLSelectElement&&o.multiple?Array.from(o.options).filter((function(e){return e.selected})).map((function(e){return e.value})):o instanceof RadioNodeList?(r=Array.from(o).filter((function(e){return e.checked})).map((function(e){return e.value}))).length>0?r[0]:"":o.value}else{var s=t.values[e.field]||"";r="boolean"==typeof s?s?"1":"":"number"==typeof s?"".concat(s):s}if("object"==typeof r)switch(e.operator){case"equals":return ae(r,[e.value]);case"notEquals":return!ae(r,[e.value]);case"contains":return null==r?void 0:r.includes(e.value);case"notContains":return!(null==r?void 0:r.includes(e.value));case"isEmpty":return null===r||0===r.length;case"isNotEmpty":return(null==r?void 0:r.length)>0;case"isOneOf":case"isNotOneOf":var l="isOneOf"===e.operator,c=(e.value?JSON.parse(e.value):[]).map((function(e){return e.toLowerCase()})),d=r.some((function(e){return c.includes(e.toLowerCase())}));return 0===c.length?l?0!==r.length:0===r.length:l?d:!d;default:return!1}switch(e.operator){case"equals":return"".concat(r).toLowerCase()==="".concat(e.value).toLowerCase();case"notEquals":return"".concat(r).toLowerCase()!=="".concat(e.value).toLowerCase();case"greaterThan":return parseFloat(r)>parseFloat(e.value);case"greaterThanOrEquals":return parseFloat(r)>=parseFloat(e.value);case"lessThan":return parseFloat(r)<parseFloat(e.value);case"lessThanOrEquals":return parseFloat(r)<=parseFloat(e.value);case"contains":return"".concat(r).toLowerCase().includes("".concat(e.value).toLowerCase());case"notContains":return!"".concat(r).toLowerCase().includes("".concat(e.value).toLowerCase());case"startsWith":return"".concat(r).toLowerCase().startsWith("".concat(e.value).toLowerCase());case"endsWith":return"".concat(r).toLowerCase().endsWith("".concat(e.value).toLowerCase());case"isEmpty":return null===r||0===r.length;case"isNotEmpty":return(null==r?void 0:r.length)>0;case"isOneOf":case"isNotOneOf":l="isOneOf"===e.operator;var u=(e.value?JSON.parse(e.value):[]).map((function(e){return e.toLowerCase()})),f=u.includes(r.toLowerCase());return 0===u.length?l?0===r.length:0!==r.length:l?f:!f;default:return!1}},this.freeform=e,this.form=e.form,this.reload()},ce=function(e){var t=this;this.reload=function(){},this.freeform=e,this.form=e.form,this.form.addEventListener(r.handleActions,(function(e){var r=e.actions.find((function(e){return"save-form"===e.name}));if(r){var n=r.metadata,o=n.key,a=n.token,s=n.url,l=u(i.saveFormhandleToken,{key:o,token:a,url:s},t.form);l.defaultPrevented||(window.location.href=l.url)}}))},de=function(){return de=Object.assign||function(e){for(var t,r=1,n=arguments.length;r<n;r++)for(var o in t=arguments[r])Object.prototype.hasOwnProperty.call(t,o)&&(e[o]=t[o]);return e},de.apply(this,arguments)},ue=function(e,t,r,n){return new(r||(r=Promise))((function(o,a){function i(e){try{l(n.next(e))}catch(e){a(e)}}function s(e){try{l(n.throw(e))}catch(e){a(e)}}function l(e){var t;e.done?o(e.value):(t=e.value,t instanceof r?t:new r((function(e){e(t)}))).then(i,s)}l((n=n.apply(e,t||[])).next())}))},fe=function(e,t){var r,n,o,a={label:0,sent:function(){if(1&o[0])throw o[1];return o[1]},trys:[],ops:[]},i=Object.create(("function"==typeof Iterator?Iterator:Object).prototype);return i.next=s(0),i.throw=s(1),i.return=s(2),"function"==typeof Symbol&&(i[Symbol.iterator]=function(){return this}),i;function s(s){return function(l){return function(s){if(r)throw new TypeError("Generator is already executing.");for(;i&&(i=0,s[0]&&(a=0)),a;)try{if(r=1,n&&(o=2&s[0]?n.return:s[0]?n.throw||((o=n.return)&&o.call(n),0):n.next)&&!(o=o.call(n,s[1])).done)return o;switch(n=0,o&&(s=[2&s[0],o.value]),s[0]){case 0:case 1:o=s;break;case 4:return a.label++,{value:s[1],done:!1};case 5:a.label++,n=s[1],s=[0];continue;case 7:s=a.ops.pop(),a.trys.pop();continue;default:if(!((o=(o=a.trys).length>0&&o[o.length-1])||6!==s[0]&&2!==s[0])){a=0;continue}if(3===s[0]&&(!o||s[1]>o[0]&&s[1]<o[3])){a.label=s[1];break}if(6===s[0]&&a.label<o[1]){a.label=o[1],o=s;break}if(o&&a.label<o[2]){a.label=o[2],a.ops.push(s);break}o[2]&&a.ops.pop(),a.trys.pop();continue}s=t.call(e,a)}catch(e){s=[6,e],n=0}finally{r=o=0}if(5&s[0])throw s[1];return{value:s[0]?s[1]:void 0,done:!0}}([s,l])}}},pe=function(){function e(n){var o=this;if(this.options={ajax:!1,disableReset:!1,disableSubmit:!1,autoScroll:!1,scrollToAnchor:!1,scrollOffset:0,scrollElement:window,showProcessingSpinner:!1,showProcessingText:!1,processingText:null,prevButtonName:"form_previous_page_button",skipHtmlReload:!1,successBannerMessage:"Form has been submitted successfully!",errorBannerMessage:"Sorry, there was an error submitting the form. Please try again.",errorClassBanner:"freeform-form-errors",errorClassList:"freeform-errors",errorClassField:"freeform-has-errors",successClassBanner:"freeform-form-success",removeMessages:null,renderSuccess:null,renderFormErrors:null,renderFieldErrors:null},this._initializedHandlers=[],this._handlers=[l,le,d,z,$,G,re,oe,D,ce],this._lockList=new Set,this._disableList=new Set,this._scrollToForm=function(){var e=o.options,t=e.scrollOffset,r=e.scrollElement,n=o.form.getBoundingClientRect().top+window.scrollY+t;r.scrollTo({top:n,behavior:o._isReducedMotion()?"instant":"smooth"})},this._isReducedMotion=function(){var e=window.matchMedia("(prefers-reduced-motion: reduce)");return!e||e.matches},this._setUp=function(){o._attachListeners(),o._getSubmitButtons().forEach((function(e){e.dataset.originalText=e.innerText,e.dataset.processingText=o.options.processingText}))},this._initHandlers=function(){o._handlers.forEach((function(e){o._initializedHandlers.push(new e(o))}))},this._resetHandlers=function(){o._initializedHandlers.forEach((function(e){return e.reload?e.reload():null}))},this.has=function(e){return null!==o.form.getAttribute(e)},this.setOption=function(e,t){o.options[e]=t},this.disableForm=function(){o.form.dataset.freeformDisabled=""},this.enableForm=function(){delete o.form.dataset.freeformDisabled},this.disableSubmit=function(e){void 0===e&&(e="freeform"),o._disableList.add(e);for(var t=0,r=Array.from(o._getSubmitButtons());t<r.length;t++)r[t].disabled=!0},this.enableSubmit=function(e){if(void 0===e&&(e="freeform"),o._disableList.delete(e),!(o._disableList.size>0))for(var t=0,r=Array.from(o._getSubmitButtons());t<r.length;t++)r[t].disabled=!1},this.lockSubmit=function(e){if(void 0===e&&(e="freeform"),o._lockList.add(e),!(o._lockList.size>1)){for(var t=o.options,r=t.disableSubmit,n=t.showProcessingSpinner,a=t.showProcessingText,i=o._getSubmitButtons(),s=0;s<i.length;s++){var l=i[s];r&&(l.disabled=!0)}var c=o._lastButtonPressed;c||(c=o._getSubmitButtons()[0]||void 0),c&&(n&&c.classList.add("freeform-processing"),a&&(c.innerText=c.dataset.processingText))}},this.unlockSubmit=function(e){void 0===e&&(e="freeform"),o._lockList.delete(e),o._lockList.size>0||o._unlockSubmitButtons()},this.forceUnlockSubmit=function(){o._lockList.clear(),o._unlockSubmitButtons()},this.triggerResubmit=function(){o.unlockSubmit(),o._lastButtonPressed?o._lastButtonPressed.click():o.triggerSubmit()},this.triggerSubmit=function(){o.unlockSubmit();var e=o._getSubmitButtons();e.length&&e[0].click()},this._unlockSubmitButtons=function(){for(var e=o.options,t=e.disableSubmit,r=e.showProcessingSpinner,n=e.showProcessingText,a=o._getSubmitButtons(),i=0;i<a.length;i++){var s=a[i];t&&(s.disabled=!1),r&&s.classList.remove("freeform-processing"),n&&(s.innerText=s.dataset.originalText)}},this._setInstances=function(){var t=o.form;e.instances.set(t,o),t.freeform=o},this._attachListeners=function(){var e=o.form,t=o.form.querySelector("input[name=freeform-action]"),n=e.querySelectorAll("[data-freeform-action]");t&&(n.forEach((function(e){return e.addEventListener("click",(function(){o._lastButtonPressed=e,t.value=e.getAttribute("data-freeform-action")}))})),e.addEventListener(r.ajaxAfterSubmit,(function(){t.value="submit"}))),e.addEventListener("submit",o._onSubmit),e.querySelectorAll("input, select, textarea").forEach((function(e){return e.addEventListener("change",(function(e){o._removeMessageFrom(e.target)}))}))},this._onSubmit=function(t){return ue(o,void 0,void 0,(function(){var n,o,a,i,s,l,c;return fe(this,(function(d){switch(d.label){case 0:if(this.lockSubmit(),t.preventDefault(),t.stopPropagation(),n=this.options.ajax,o=t.submitter,a=!1,o&&o.name&&o.name===e._BACK_BUTTON_NAME&&(a=!0),i={},this._dispatchEvent(r.submit,{isBackButtonPressed:a,cancelable:!0,addCallback:function(e,t){void 0===t&&(t=0),void 0===i[t]&&(i[t]=[]),i[t].push(e)}}).defaultPrevented)return this.forceUnlockSubmit(),this._dispatchEvent(r.afterFailedSubmit,{cancelable:!1}),[2,!1];s=Object.entries(i).sort((function(e,t){var r=e[0],n=t[0];return Number(r)-Number(n)})).flatMap((function(e){return e[1]})),l=0,c=s,d.label=1;case 1:return l<c.length?[4,(0,c[l])()]:[3,4];case 2:if(!1===d.sent())return this.forceUnlockSubmit(),this._dispatchEvent(r.afterFailedSubmit,{cancelable:!1}),[2,!1];d.label=3;case 3:return l++,[3,1];case 4:return n?(this._onSubmitAjax(t),[2,!1]):(this.form.submit(),[2])}}))}))},this._removeMessages=function(){if(!o._dispatchEvent(r.removeMessages).defaultPrevented){if("function"==typeof o.options.removeMessages)return o.options.removeMessages=o.options.removeMessages.bind(o),o.options.removeMessages();var e=o,t=e.form,n=e.options,a=n.successClassBanner,i=n.errorClassBanner,s=n.errorClassList,l=n.errorClassField;Q(t.querySelectorAll(".".concat(K(s).join(".")))),t.querySelectorAll(".".concat(K(l).join("."))).forEach((function(e){o._removeMessageFrom(e)})),Q(t.querySelectorAll(P(a))),Q(document.querySelectorAll(P(i)))}},this._removeMessageFrom=function(e){if(!o._dispatchEvent(r.fieldRemoveMessages,{field:e}).defaultPrevented){var t=o.options,n=t.errorClassList,a=t.errorClassField,i=e.parentNode;e.type&&("radio"===e.type||"checkbox"===e.type&&/\[]$/.test(e.name))&&(i=e.parentNode.parentNode),Q(i.querySelector(P(n)));for(var s=i.querySelectorAll("input, select, textarea"),l=0;l<s.length;l++)Z(s[l],a)}},this._renderSuccessBanner=function(){if(!o._dispatchEvent(r.renderSuccess).defaultPrevented){if("function"==typeof o.options.renderSuccess)return o.options.renderSuccess=o.options.renderSuccess.bind(o),o.options.renderSuccess();var e=o,t=e.form,n=e.options,a=n.successBannerMessage,i=n.successClassBanner,s=document.createElement("div");Y(s,i);var l=document.createElement("p");l.appendChild(document.createTextNode(a)),s.appendChild(l),t.insertBefore(s,t.childNodes[0])}},this._renderFieldErrors=function(e){if(o._dispatchEvent(r.renderFieldErrors,{errors:e}).defaultPrevented)return!1;if("function"==typeof o.options.renderFieldErrors)return o.options.renderFieldErrors=o.options.renderFieldErrors.bind(o),o.options.renderFieldErrors(e);var t=o,n=t.form,a=t.options,i=a.errorClassList,s=a.errorClassField;for(var l in e){var c=e[l],d=document.createElement("ul");d.setAttribute("data-field-errors",""),Y(d,i);for(var u=0;u<c.length;u++){var f=c[u],p=document.createElement("li");p.appendChild(document.createTextNode(f)),d.appendChild(p)}var m=n.querySelector('[data-field-container="'.concat(l,'"]')),h=n.querySelector('[data-error-append-target="'.concat(l,'"]')),v=n.querySelectorAll('\n          [name="'.concat(l,'"],\n          [type=file][name="').concat(l,'"],\n          [type=file][name="').concat(l,'[]"],\n          [type=checkbox][name="').concat(l,'[]"],\n          [type=radio][name="').concat(l,'"],\n          select[multiple][name="').concat(l,'[]"],\n          [data-freeform-file-upload="').concat(l,'"]\n        '));if(!m)return;for(var b=0;b<v.length;b++){var g=v[b];Y(g,s)}h?h.appendChild(d):m.appendChild(d)}},this._renderFormErrors=function(e){if(o._dispatchEvent(r.renderFormErrors,{errors:e}).defaultPrevented)return!1;if("function"==typeof o.options.renderFormErrors)return o.options.renderFormErrors=o.options.renderFormErrors.bind(o),o.options.renderFormErrors(e);var t=o,n=t.form,a=t.options,i=a.errorClassBanner,s=a.errorBannerMessage,l=document.createElement("div");Y(l,i);var c=document.createElement("p");if(c.appendChild(document.createTextNode(s)),l.appendChild(c),e.length){for(var d=document.createElement("ul"),u=0;u<e.length;u++){var f=e[u],p=document.createElement("li");p.appendChild(document.createTextNode(f)),d.appendChild(p)}l.appendChild(d)}n.insertBefore(l,n.childNodes[0])},this._prepareFormData=function(){var e=o.form,t=new FormData(e);if(navigator.userAgent.indexOf("Safari")>-1)for(var r=0;r<e.elements.length;r++){var n=e.elements[r];"file"===n.type&&""===n.value&&t.delete(n.name)}return t},this.quickSave=function(e,t){return ue(o,void 0,void 0,(function(){var n,o,a,i,s,l,c,d,u,f;return fe(this,(function(p){switch(p.label){case 0:n=this.form,(o=this._prepareFormData()).set("action","freeform/submit/quick-save"),o.set("storage-secret",e),t&&o.set("token",t),p.label=1;case 1:return p.trys.push([1,3,,4]),[4,E(n.getAttribute("action")||window.location.href,{method:n.getAttribute("method"),data:o})];case 2:return a=p.sent(),[3,4];case 3:return i=p.sent(),417===(null===(f=null==i?void 0:i.response)||void 0===f?void 0:f.status)?(this.unlockSubmit(),[2,!1]):[3,4];case 4:if(this._removeMessages(),s=a.data,200===a.status){if(l=s.success,c=s.errors,d=s.formErrors,u=s.storageToken,l)return[2,u];(c||d)&&(this._dispatchEvent(r.ajaxError,{request:a,response:s,errors:c,formErrors:d}),this._dispatchEvent(r.afterFailedSubmit,{cancelable:!1}),this._renderFieldErrors(c),this._renderFormErrors(d)),this.options.autoScroll&&this._scrollToForm()}else this._dispatchEvent(r.ajaxError,{request:a,response:s}),this._dispatchEvent(r.afterFailedSubmit,{cancelable:!1});return this.unlockSubmit(),[2]}}))}))},this._onSubmitAjax=function(e){var n=o.form,a=o._prepareFormData(),i=new XMLHttpRequest,s=e.submitter;s&&s.name&&a.append(s.name,"1");var l=n.getAttribute("method"),c=n.getAttribute("action");i.open(l,c||window.location.href,!0),i.setRequestHeader("Cache-Control","no-cache"),i.setRequestHeader("X-Requested-With","XMLHttpRequest"),i.setRequestHeader("HTTP_X_REQUESTED_WITH","XMLHttpRequest"),i.onload=function(){if(o._removeMessages(),200===i.status){var e=(h=JSON.parse(i.response)).success,s=h.finished,l=h.actions,c=void 0===l?[]:l,d=h.errors,u=h.formErrors,f=h.returnUrl;if(o._dispatchEvent(r.ajaxBeforeSuccess,{request:i,response:h}).defaultPrevented)return;if(c.length)o._dispatchEvent(r.handleActions,{response:h,actions:c,cancelable:!1});else if(e){if(s&&h.onSuccess===t.RedirectReturnUrl&&f){if(o._dispatchEvent(r.ajaxSuccess,{request:i,response:h}).defaultPrevented)return;return void(window.location.href=f)}null===h.html||o.options.skipHtmlReload||(n.innerHTML=h.html.replace(/<form[^>]*>/,"").replace("</form>","")),o.options.skipHtmlReload||(o._resetHandlers(),o._setUp()),s&&(o.options.disableReset||(n.reset(),o._dispatchEvent(r.reset)),h.onSuccess===t.Reload&&o._renderSuccessBanner()),o._dispatchEvent(r.ajaxSuccess,{request:i,response:h})}else(d||u)&&(o._dispatchEvent(r.ajaxError,{request:i,response:h,errors:d,formErrors:u}),o._dispatchEvent(r.afterFailedSubmit,{cancelable:!1}),o._renderFieldErrors(d),o._renderFormErrors(u));var p=null==h?void 0:h.freeform_payload;if(p){var m=n.querySelector("input[name^=freeform_payload]");m&&(m.value=p)}o._dispatchEvent(r.ajaxAfterSubmit,{data:a,request:i,response:h,cancelable:!1}),o.options.autoScroll&&o._scrollToForm()}else{var h=i.response;o._dispatchEvent(r.ajaxError,{request:i,response:h})}o.unlockSubmit()},o._dispatchEvent(r.ajaxBeforeSubmit,{data:a,request:i}).defaultPrevented||i.send(a)},this._getSubmitButtons=function(){var e=o.form.querySelectorAll("*[type=submit][data-freeform-action]");return e.length?e:o.form.querySelectorAll("*[type=submit]")},this._getBackButtons=function(){return o.form.querySelectorAll('*[type=submit][data-freeform-action="back"]')},this._dispatchEvent=function(e,t,r){var n=u(e,de(de({},t),{form:o.form,freeform:o}),r);return document.dispatchEvent(n),o.form.dispatchEvent(n),n},e.instances.get(n))return e.instances.get(n);this.id=n.dataset.id,this.form=n,this._setInstances();var a={ajax:null!==n.getAttribute("data-ajax"),disableReset:null!==n.getAttribute("data-disable-reset"),scrollToAnchor:null!==n.getAttribute("data-scroll-to-anchor"),autoScroll:null!==n.getAttribute("data-auto-scroll"),disableSubmit:null!==n.getAttribute("data-disable-submit"),showProcessingSpinner:null!==n.getAttribute("data-show-processing-spinner"),showProcessingText:null!==n.getAttribute("data-show-processing-text"),processingText:n.getAttribute("data-processing-text"),successBannerMessage:n.getAttribute("data-success-message"),errorBannerMessage:n.getAttribute("data-error-message"),skipHtmlReload:null!==n.getAttribute("data-skip-html-reload")};this.options=de(de({},this.options),a),this.disableSubmit("init");var i=setInterval((function(){return ue(o,void 0,void 0,(function(){var e;return fe(this,(function(t){return"complete"===document.readyState&&(clearInterval(i),e=this._dispatchEvent(r.ready,{options:{}}),this.options=de(de({},this.options),e.options),this._setUp(),this._initHandlers(),this.enableSubmit("init"),this.options.scrollToAnchor&&this._scrollToForm()),[2]}))}))}),50)}return e._BACK_BUTTON_NAME="form_previous_page_button",e.instances=new WeakMap,e.getInstance=function(t){return e.instances.get(t)},e}(),me=pe;document.querySelectorAll("form[data-freeform]").forEach((function(e){new pe(e)}));var he,ve=function(e){var t;"FORM"===e.nodeName&&void 0!==(null===(t=e.dataset)||void 0===t?void 0:t.freeform)&&new pe(e),null==e||e.childNodes.forEach(ve)},be=new MutationObserver((function(e){e.forEach((function(e){"childList"===e.type&&e.addedNodes.forEach((function(e){ve(e)}))}))})),ge=0,ye=function(){if(ge>25)return console.warn("Freeform observer timed out"),clearTimeout(he);document.body?be.observe(document.body,{childList:!0,subtree:!0}):(ge++,he&&clearTimeout(he),he=setTimeout(ye,50))};ye()}();
+!function(){"use strict";var e={d:function(t,r){for(var n in r)e.o(r,n)&&!e.o(t,n)&&Object.defineProperty(t,n,{enumerable:!0,get:r[n]})},o:function(e,t){return Object.prototype.hasOwnProperty.call(e,t)}};e.d({},{A:function(){return ve}});var t,r={ready:"freeform-ready",reset:"freeform-on-reset",submit:"freeform-on-submit",removeMessages:"freeform-remove-messages",fieldRemoveMessages:"freeform-remove-field-messages",renderSuccess:"freeform-render-success",renderFieldErrors:"freeform-render-field-errors",renderFormErrors:"freeform-render-form-errors",ajaxBeforeSuccess:"freeform-before-ajax-success",ajaxSuccess:"freeform-ajax-success",ajaxError:"freeform-ajax-error",ajaxBeforeSubmit:"freeform-ajax-before-submit",ajaxAfterSubmit:"freeform-ajax-after-submit",afterFailedSubmit:"freeform-after-failed-submit",handleActions:"freeform-handle-actions"},n={applied:"freeform-rules-applied"},o={onAddRow:"freeform-field-table-on-add-row",afterRowAdded:"freeform-field-table-after-row-added",onRemoveRow:"freeform-field-table-on-remove-row",afterRemoveRow:"freeform-field-table-after-remove-row"},a={renderPreview:"freeform-field-dnd-on-render-preview",renderPreviewRemoveButton:"freeform-field-dnd-on-render-preview-remove-button",renderErrorContainer:"freeform-field-dnd-render-error-container",showGlobalMessage:"freeform-field-dnd-show-global-message",appendErrors:"freeform-field-dnd-append-errors",clearErrors:"freeform-field-dnd-clear-errors",onChange:"freeform-field-dnd-on-change",onUploadProgress:"freeform-field-dnd-on-upload-progress"},i={saveFormhandleToken:"freeform-save-form-handle-token"};!function(e){e.Reload="reload",e.RedirectReturnUrl="redirect-return-url",e.LoadSuccessTemplate="load-success-template"}(t||(t={}));var s,l,c=function(e){var t=this;this.reload=function(){for(var e=t.form.querySelectorAll("*[name=".concat(ve._BACK_BUTTON_NAME,"]")),r=0;r<e.length;r++)e[r].addEventListener("click",(function(){var e=document.createElement("input");e.type="hidden",e.name=ve._BACK_BUTTON_NAME,e.value="",t.form.appendChild(e)}))},this.form=e.form,this.reload()},u=function(){return u=Object.assign||function(e){for(var t,r=1,n=arguments.length;r<n;r++)for(var o in t=arguments[r])Object.prototype.hasOwnProperty.call(t,o)&&(e[o]=t[o]);return e},u.apply(this,arguments)},d=function(e){var t=this;if(this.loadedLocales={},this.scriptAdded=!1,this.reload=function(){t.freeform.has("data-scripts-datepicker")&&t.freeform.form.querySelectorAll("*[data-datepicker][data-datepicker-enabled]").forEach((function(e){var r=e.getAttribute("data-datepicker-locale"),n={disableMobile:!0,allowInput:!0,dateFormat:e.getAttribute("data-datepicker-format"),enableTime:null!==e.getAttribute("data-datepicker-enabletime"),noCalendar:null===e.getAttribute("data-datepicker-enabledate"),time_24hr:null!==e.getAttribute("data-datepicker-clock_24h"),minDate:e.getAttribute("data-datepicker-min-date"),maxDate:e.getAttribute("data-datepicker-max-date"),minuteIncrement:1,hourIncrement:1,static:null!==e.getAttribute("data-datepicker-static")},o=t.freeform._dispatchEvent("flatpickr-before-init",{detail:n,options:n}),a=u(u({},o.detail),o.options),i=flatpickr(e,a);if(e.setAttribute("autocomplete","off"),t.freeform._dispatchEvent("flatpickr-ready",{detail:i,flatpickr:i}),t.loadedLocales[r])t.loadedLocales[r].addEventListener("load",(function(){i.set("locale",r),t.loadedLocales[r].dataset.loaded="true"})),"true"===t.loadedLocales[r].dataset.loaded&&i.set("locale",r);else{var s=document.createElement("script");s.src="//cdnjs.cloudflare.com/ajax/libs/flatpickr/4.6.13/l10n/".concat(r,".js"),s.async=!1,s.defer=!1,s.dataset.loaded="",s.addEventListener("load",(function(){i.set("locale",r),s.dataset.loaded="true"})),document.body.appendChild(s),t.loadedLocales[r]=s}}))},this.freeform=e,this.freeform.has("data-scripts-datepicker")&&!this.scriptAdded){var r=document.createElement("script");r.src="//cdnjs.cloudflare.com/ajax/libs/flatpickr/4.6.13/flatpickr.min.js",r.async=!1,r.defer=!1,r.addEventListener("load",(function(){t.reload()})),document.body.appendChild(r);var n=document.createElement("link");n.rel="stylesheet",n.href="//cdnjs.cloudflare.com/ajax/libs/flatpickr/4.6.13/flatpickr.min.css",document.body.appendChild(n),this.scriptAdded=!0}},f=function(e,t,r){var n=t||{},o=n.bubbles,a=void 0!==o&&o,i=n.cancelable,s=void 0===i||i,l=function(e,t){var r={};for(var n in e)Object.prototype.hasOwnProperty.call(e,n)&&t.indexOf(n)<0&&(r[n]=e[n]);if(null!=e&&"function"==typeof Object.getOwnPropertySymbols){var o=0;for(n=Object.getOwnPropertySymbols(e);o<n.length;o++)t.indexOf(n[o])<0&&Object.prototype.propertyIsEnumerable.call(e,n[o])&&(r[n[o]]=e[n[o]])}return r}(n,["bubbles","cancelable"]),c=p(e,a,s);return Object.assign(c,l),r&&(r instanceof HTMLElement?r.dispatchEvent(c):Array.from(r).forEach((function(e){return e.dispatchEvent(c)}))),c},p=function(e,t,r){return void 0===t&&(t=!0),void 0===r&&(r=!0),new Event(e,{bubbles:t,cancelable:r})},m=function(){return e=void 0,t=void 0,n=function(){var e,t,r;return function(e,t){var r,n,o,a={label:0,sent:function(){if(1&o[0])throw o[1];return o[1]},trys:[],ops:[]},i=Object.create(("function"==typeof Iterator?Iterator:Object).prototype);return i.next=s(0),i.throw=s(1),i.return=s(2),"function"==typeof Symbol&&(i[Symbol.iterator]=function(){return this}),i;function s(s){return function(l){return function(s){if(r)throw new TypeError("Generator is already executing.");for(;i&&(i=0,s[0]&&(a=0)),a;)try{if(r=1,n&&(o=2&s[0]?n.return:s[0]?n.throw||((o=n.return)&&o.call(n),0):n.next)&&!(o=o.call(n,s[1])).done)return o;switch(n=0,o&&(s=[2&s[0],o.value]),s[0]){case 0:case 1:o=s;break;case 4:return a.label++,{value:s[1],done:!1};case 5:a.label++,n=s[1],s=[0];continue;case 7:s=a.ops.pop(),a.trys.pop();continue;default:if(!((o=(o=a.trys).length>0&&o[o.length-1])||6!==s[0]&&2!==s[0])){a=0;continue}if(3===s[0]&&(!o||s[1]>o[0]&&s[1]<o[3])){a.label=s[1];break}if(6===s[0]&&a.label<o[1]){a.label=o[1],o=s;break}if(o&&a.label<o[2]){a.label=o[2],a.ops.push(s);break}o[2]&&a.ops.pop(),a.trys.pop();continue}s=t.call(e,a)}catch(e){s=[6,e],n=0}finally{r=o=0}if(5&s[0])throw s[1];return{value:s[0]?s[1]:void 0,done:!0}}([s,l])}}}(this,(function(n){switch(n.label){case 0:if(void 0!==s)return[3,5];n.label=1;case 1:return n.trys.push([1,4,,5]),e=document.querySelector("form[data-csrf-info]"),t=void 0,e&&(t=e.dataset.csrfInfo),t||(t="/index.php?p=actions/users/session-info"),[4,fetch(t,{headers:{Accept:"application/json"}})];case 2:return[4,n.sent().json()];case 3:return r=n.sent(),s=void 0!==r.csrfTokenName?{name:r.csrfTokenName,value:r.csrfTokenValue}:null,[3,5];case 4:return n.sent(),s=null,[3,5];case 5:return[2,s]}}))},new((r=void 0)||(r=Promise))((function(o,a){function i(e){try{l(n.next(e))}catch(e){a(e)}}function s(e){try{l(n.throw(e))}catch(e){a(e)}}function l(e){var t;e.done?o(e.value):(t=e.value,t instanceof r?t:new r((function(e){e(t)}))).then(i,s)}l((n=n.apply(e,t||[])).next())}));var e,t,r,n},h=(l=function(e,t){return l=Object.setPrototypeOf||{__proto__:[]}instanceof Array&&function(e,t){e.__proto__=t}||function(e,t){for(var r in t)Object.prototype.hasOwnProperty.call(t,r)&&(e[r]=t[r])},l(e,t)},function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Class extends value "+String(t)+" is not a constructor or null");function r(){this.constructor=e}l(e,t),e.prototype=null===t?Object.create(t):(r.prototype=t.prototype,new r)}),v=function(){return v=Object.assign||function(e){for(var t,r=1,n=arguments.length;r<n;r++)for(var o in t=arguments[r])Object.prototype.hasOwnProperty.call(t,o)&&(e[o]=t[o]);return e},v.apply(this,arguments)},b=function(){function e(){}return e.prototype.cancel=function(){this.cancelFn&&(this.cancelFn(),this.cancelFn=null)},e.prototype._setCancelFn=function(e){this.cancelFn=e},e}(),y=function(e){function t(r,n,o){var a=e.call(this,r)||this;return a.response=v(v({},n),{data:o}),a.status=n.status,Object.setPrototypeOf(a,t.prototype),a}return h(t,e),t}(Error),g=function(e,t,r,n,o){return a=void 0,i=void 0,l=function(){var a,i;return function(e,t){var r,n,o,a={label:0,sent:function(){if(1&o[0])throw o[1];return o[1]},trys:[],ops:[]},i=Object.create(("function"==typeof Iterator?Iterator:Object).prototype);return i.next=s(0),i.throw=s(1),i.return=s(2),"function"==typeof Symbol&&(i[Symbol.iterator]=function(){return this}),i;function s(s){return function(l){return function(s){if(r)throw new TypeError("Generator is already executing.");for(;i&&(i=0,s[0]&&(a=0)),a;)try{if(r=1,n&&(o=2&s[0]?n.return:s[0]?n.throw||((o=n.return)&&o.call(n),0):n.next)&&!(o=o.call(n,s[1])).done)return o;switch(n=0,o&&(s=[2&s[0],o.value]),s[0]){case 0:case 1:o=s;break;case 4:return a.label++,{value:s[1],done:!1};case 5:a.label++,n=s[1],s=[0];continue;case 7:s=a.ops.pop(),a.trys.pop();continue;default:if(!((o=(o=a.trys).length>0&&o[o.length-1])||6!==s[0]&&2!==s[0])){a=0;continue}if(3===s[0]&&(!o||s[1]>o[0]&&s[1]<o[3])){a.label=s[1];break}if(6===s[0]&&a.label<o[1]){a.label=o[1],o=s;break}if(o&&a.label<o[2]){a.label=o[2],a.ops.push(s);break}o[2]&&a.ops.pop(),a.trys.pop();continue}s=t.call(e,a)}catch(e){s=[6,e],n=0}finally{r=o=0}if(5&s[0])throw s[1];return{value:s[0]?s[1]:void 0,done:!0}}([s,l])}}}(this,(function(s){switch(s.label){case 0:return(a=o.request||new XMLHttpRequest).open(e,t),a.setRequestHeader("Cache-Control","no-cache"),a.setRequestHeader("X-Requested-With","XMLHttpRequest"),a.setRequestHeader("HTTP_X_REQUESTED_WITH","XMLHttpRequest"),[4,m()];case 1:return(i=s.sent())&&a.setRequestHeader("X-CSRF-Token",i.value),w(a,null==o?void 0:o.headers),a.onload=function(){var e=a.response;try{e=JSON.parse(a.response)}catch(e){}var t=a.status;t<200||t>=300?n(new y("Request failed with status ".concat(a.statusText),a,e)):r({status:a.status,data:e,request:a})},a.onerror=function(){n(new Error("Network error"))},a.onabort=function(){n(new Error("Request aborted"))},o.onUploadProgress&&(a.upload.onprogress=function(e){o.onUploadProgress(e)}),o.cancelToken&&o.cancelToken._setCancelFn((function(){a.abort()})),[2,a]}}))},new((s=void 0)||(s=Promise))((function(e,t){function r(e){try{o(l.next(e))}catch(e){t(e)}}function n(e){try{o(l.throw(e))}catch(e){t(e)}}function o(t){var o;t.done?e(t.value):(o=t.value,o instanceof s?o:new s((function(e){e(o)}))).then(r,n)}o((l=l.apply(a,i||[])).next())}));var a,i,s,l},w=function(e,t){t&&Object.entries(t).forEach((function(t){var r=t[0],n=t[1];e.setRequestHeader(r,String(n))}))},E=function(e,t,r,n){return new(r||(r=Promise))((function(o,a){function i(e){try{l(n.next(e))}catch(e){a(e)}}function s(e){try{l(n.throw(e))}catch(e){a(e)}}function l(e){var t;e.done?o(e.value):(t=e.value,t instanceof r?t:new r((function(e){e(t)}))).then(i,s)}l((n=n.apply(e,t||[])).next())}))},S=function(e,t){var r,n,o,a={label:0,sent:function(){if(1&o[0])throw o[1];return o[1]},trys:[],ops:[]},i=Object.create(("function"==typeof Iterator?Iterator:Object).prototype);return i.next=s(0),i.throw=s(1),i.return=s(2),"function"==typeof Symbol&&(i[Symbol.iterator]=function(){return this}),i;function s(s){return function(l){return function(s){if(r)throw new TypeError("Generator is already executing.");for(;i&&(i=0,s[0]&&(a=0)),a;)try{if(r=1,n&&(o=2&s[0]?n.return:s[0]?n.throw||((o=n.return)&&o.call(n),0):n.next)&&!(o=o.call(n,s[1])).done)return o;switch(n=0,o&&(s=[2&s[0],o.value]),s[0]){case 0:case 1:o=s;break;case 4:return a.label++,{value:s[1],done:!1};case 5:a.label++,n=s[1],s=[0];continue;case 7:s=a.ops.pop(),a.trys.pop();continue;default:if(!((o=(o=a.trys).length>0&&o[o.length-1])||6!==s[0]&&2!==s[0])){a=0;continue}if(3===s[0]&&(!o||s[1]>o[0]&&s[1]<o[3])){a.label=s[1];break}if(6===s[0]&&a.label<o[1]){a.label=o[1],o=s;break}if(o&&a.label<o[2]){a.label=o[2],a.ops.push(s);break}o[2]&&a.ops.pop(),a.trys.pop();continue}s=t.call(e,a)}catch(e){s=[6,e],n=0}finally{r=o=0}if(5&s[0])throw s[1];return{value:s[0]?s[1]:void 0,done:!0}}([s,l])}}},k=function(e,t,r){if(r||2===arguments.length)for(var n,o=0,a=t.length;o<a;o++)!n&&o in t||(n||(n=Array.prototype.slice.call(t,0,o)),n[o]=t[o]);return e.concat(n||Array.prototype.slice.call(t))},A=function(e,t){return new Promise((function(r,n){g((null==t?void 0:t.method)||"GET",e,r,n,t).then((function(e){var r=null==t?void 0:t.data;r instanceof FormData?e.send(r):(e.setRequestHeader("Content-Type","application/json"),e.send(JSON.stringify(r)))}))}))};A.get=function(e){for(var t=[],r=1;r<arguments.length;r++)t[r-1]=arguments[r];return E(void 0,k([e],t,!0),void 0,(function(e,t){return void 0===t&&(t={}),S(this,(function(r){return[2,new Promise((function(r,n){g("GET",e,r,n,t).then((function(t){t.open("GET",e),t.send()}))}))]}))}))},A.post=function(e,t){for(var r=[],n=2;n<arguments.length;n++)r[n-2]=arguments[n];return E(void 0,k([e,t],r,!0),void 0,(function(e,t,r){return void 0===r&&(r={}),S(this,(function(n){return[2,new Promise((function(n,o){g("POST",e,n,o,r).then((function(e){t instanceof FormData?e.send(t):(e.setRequestHeader("Content-Type","application/json"),e.send(JSON.stringify(t)))}))}))]}))}))};const _="bits",x="bytes",L="jedec",C={symbol:{iec:{bits:["bit","Kibit","Mibit","Gibit","Tibit","Pibit","Eibit","Zibit","Yibit"],bytes:["B","KiB","MiB","GiB","TiB","PiB","EiB","ZiB","YiB"]},jedec:{bits:["bit","Kbit","Mbit","Gbit","Tbit","Pbit","Ebit","Zbit","Ybit"],bytes:["B","KB","MB","GB","TB","PB","EB","ZB","YB"]}},fullform:{iec:["","kibi","mebi","gibi","tebi","pebi","exbi","zebi","yobi"],jedec:["","kilo","mega","giga","tera","peta","exa","zetta","yotta"]}};var T,j=function(e,t,r){if(r||2===arguments.length)for(var n,o=0,a=t.length;o<a;o++)!n&&o in t||(n||(n=Array.prototype.slice.call(t,0,o)),n[o]=t[o]);return e.concat(n||Array.prototype.slice.call(t))},q="freeform-file-dnd",F=function(e){for(var t=[],r=1;r<arguments.length;r++)t[r-1]=arguments[r];e.classList.add(t.join("__"))},O=function(e){for(var t=[],r=1;r<arguments.length;r++)t[r-1]=arguments[r];e.classList.remove(t.join("__"))},P=function(e){for(var t=[],r=1;r<arguments.length;r++)t[r-1]=arguments[r];F.apply(void 0,j([e,q],t,!1))},B=function(e){return"string"==typeof e&&(e=e.split(" ")),e.map((function(e){return".".concat(e)})).join("")},R=function(e,t){var r=function(e){var t=e.name,r=e.extension,n=e.size,o=document.createElement("div");o.setAttribute("data-file-preview",""),P(o,"preview-zone","file-preview");var a=document.createElement("span");a.setAttribute("data-extension-label",""),a.innerText=r.toUpperCase(),P(a,"preview-zone","file-preview","thumbnail","extension-label");var i=document.createElement("div");i.setAttribute("data-thumbnail",""),i.appendChild(a),P(i,"preview-zone","file-preview","thumbnail");var s,l,c,u,d=document.createElement("span");d.setAttribute("data-filename",""),d.innerText=(void 0===(l=14)&&(l=50),void 0===c&&(c="…"),void 0===u&&(u=3),(s=t).length>l?s.substr(0,l-u)+c:s),d.title=t,P(d,"preview-zone","file-preview","filename");var f=document.createElement("span");return f.setAttribute("data-filesize",""),f.innerText=n,P(f,"preview-zone","file-preview","filesize"),o.appendChild(i),o.appendChild(d),o.appendChild(f),[o,i,d,f]}(e),n=r[0],o=r[1];e.url&&(o.style.backgroundImage="url(".concat(e.url,")"));var i=function(e){var t=document.createElement("a");return t.setAttribute("data-remove-button",""),t.setAttribute("title","Remove file?"),t.setAttribute("aria-label","Remove file?"),t.innerHTML='<svg style="height: 14px; width: 14px;" aria-hidden="true" focusable="false" data-prefix="fas" data-icon="times" class="svg-inline--fa fa-times fa-w-11" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 352 512"><path fill="currentColor" d="M242.72 256l100.07-100.07c12.28-12.28 12.28-32.19 0-44.48l-22.24-22.24c-12.28-12.28-32.19-12.28-44.48 0L176 189.28 75.93 89.21c-12.28-12.28-32.19-12.28-44.48 0L9.21 111.45c-12.28 12.28-12.28 32.19 0 44.48L109.28 256 9.21 356.07c-12.28 12.28-12.28 32.19 0 44.48l22.24 22.24c12.28 12.28 32.2 12.28 44.48 0L176 322.72l100.07 100.07c12.28 12.28 32.2 12.28 44.48 0l22.24-22.24c12.28-12.28 12.28-32.19 0-44.48L242.72 256z"></path></svg>',P(t,"preview-zone","file-preview","thumbnail","remove-button"),e._dispatchEvent(a.renderPreviewRemoveButton,{button:t}).button}(t),s=function(e){var t=document.createElement("div");return t.innerText="!",t.setAttribute("data-errors",""),P(t,"preview-zone","file-preview","thumbnail","errors"),e._dispatchEvent(a.renderErrorContainer,{container:t},t).container}(t),l=function(){var e=document.createElement("div");return e.setAttribute("data-progress",""),P(e,"preview-zone","file-preview","thumbnail","progress"),e}();o.appendChild(i),o.appendChild(s),o.appendChild(l);var c=t._dispatchEvent(a.renderPreview,{metadata:e,container:n});return P(n,"preview-zone","file-preview","animation-enter"),setTimeout((function(){!function(e){for(var t=[],r=1;r<arguments.length;r++)t[r-1]=arguments[r];O.apply(void 0,j([e,q],t,!1))}(n,"preview-zone","file-preview","animation-enter")}),10),c.container.addEventListener("click",(function(e){return e.stopPropagation()})),c.container},M=function(e,t){var r=t.id,n=document.createElement("input");return n.type="hidden",n.value=r,n.name="".concat(e,"[]"),n};!function(e){e.FieldError="field-error"}(T||(T={}));var N=["png","jpg","jpeg","gif"],I=function(e){return N.includes(e.toLowerCase())},U=function(e,t,r,n,o){var i=new b,s=function(){i.cancel()},l=e.name.match(/.(\w+)$/i),c=e.name,u=function(e,{bits:t=!1,pad:r=!1,base:n=-1,round:o=2,locale:a="",localeOptions:i={},separator:s="",spacer:l=" ",symbols:c={},standard:u="",output:d="string",fullform:f=!1,fullforms:p=[],exponent:m=-1,roundingMethod:h="round",precision:v=0}={}){let b=m,y=Number(e),g=[],w=0,E="";"si"===u?(n=10,u=L):"iec"===u||u===L?n=2:2===n?u="iec":(n=10,u=L);const S=10===n?1e3:1024,k=!0===f,A=y<0,T=Math[h];if("bigint"!=typeof e&&isNaN(e))throw new TypeError("Invalid number");if("function"!=typeof T)throw new TypeError("Invalid rounding method");if(A&&(y=-y),(-1===b||isNaN(b))&&(b=Math.floor(Math.log(y)/Math.log(S)),b<0&&(b=0)),b>8&&(v>0&&(v+=8-b),b=8),"exponent"===d)return b;if(0===y)g[0]=0,E=g[1]=C.symbol[u][t?_:x][b];else{w=y/(2===n?Math.pow(2,10*b):Math.pow(1e3,b)),t&&(w*=8,w>=S&&b<8&&(w/=S,b++));const e=Math.pow(10,b>0?o:0);g[0]=T(w*e)/e,g[0]===S&&b<8&&-1===m&&(g[0]=1,b++),E=g[1]=10===n&&1===b?t?"kbit":"kB":C.symbol[u][t?_:x][b]}if(A&&(g[0]=-g[0]),v>0&&(g[0]=g[0].toPrecision(v)),g[1]=c[g[1]]||g[1],!0===a?g[0]=g[0].toLocaleString():a.length>0?g[0]=g[0].toLocaleString(a,i):s.length>0&&(g[0]=g[0].toString().replace(".",s)),r&&o>0){const e=g[0].toString(),t=s||(e.match(/(\D)/g)||[]).pop()||".",r=e.toString().split(t),n=r[1]||"",a=n.length,i=o-a;g[0]=`${r[0]}${t}${n.padEnd(a+i,"0")}`}return k&&(g[1]=p[b]?p[b]:C.fullform[u][b]+(t?"bit":"byte")+(1===g[0]?"":"s")),"array"===d?g:"object"===d?{value:g[0],symbol:g[1],exponent:b,unit:E}:g.join(l)}(e.size,{round:1}),d=null!==l?l[1].toLowerCase():"n/a",p=R({name:c,extension:d,size:u},o),m=p.querySelector("[data-thumbnail]"),h=p.querySelector("[data-remove-button]");if(I(d)){var v=new FileReader;v.readAsDataURL(e),v.onloadend=function(){m.setAttribute("data-has-preview",""),m.style.backgroundImage="url(".concat(v.result.toString(),")")}}n.appendChild(p),h.addEventListener("click",s),D(r);var y=new FormData(o.form);y.delete("action"),y.append("handle",t),y.append(t,e);var g=r.getAttribute("data-base-url");return A.post("".concat(g,"/files/upload"),y,{cancelToken:i,onUploadProgress:function(e){var t=e.total,n=e.loaded,o=Math.ceil(n/(t/100));f(a.onUploadProgress,{total:t,loaded:n,percent:o},r),p.style.setProperty("--progress","".concat(o,"%")),o>=98&&h.removeEventListener("click",s)}}).then((function(e){var a=new FormData(o.form);a.delete("action"),a.append("handle",t),a.append("id",e.data.id),h.removeEventListener("click",s),h.addEventListener("click",(function(){confirm("Are you sure?")&&A.post("".concat(g,"/files/delete"),a).then((function(){n.removeChild(p),D(r)})).catch((function(e){alert(e.message)}))})),p.appendChild(M(t,e.data)),p.setAttribute("data-completed","")})).catch((function(e){var t,i,l;if("Request aborted"===e.message)return n.removeChild(p),void D(r);if(h.removeEventListener("click",s),h.addEventListener("click",(function(){n.removeChild(p),D(r)})),(null===(i=null===(t=null==e?void 0:e.response)||void 0===t?void 0:t.data)||void 0===i?void 0:i.type)===T.FieldError){var c=(null===(l=null==e?void 0:e.response)||void 0===l?void 0:l.data).messages;!function(e,t,r,n){if(!n._dispatchEvent(a.clearErrors,{},e).defaultPrevented){var o=t.querySelector("[data-errors]");t.setAttribute("data-has-errors",""),o.setAttribute("aria-label",r.join("; ")),o.setAttribute("title",r.join(". "))}}(r,p,c,o)}else console.warn(e)}))},D=function(e){f(a.onChange,{container:e},e)},H=function(e,t,r){void 0===r&&(r=4e3);var n=document.createElement("li");n.setAttribute("data-error",""),n.innerText=t,P(n,"messages","message"),P(n,"messages","message","error");var o=f(a.showGlobalMessage,{messageItem:n},e);z(e,o.messageItem,r)},z=function(e,t,r){var n=e.querySelector("[data-messages]");n&&(n.appendChild(t),setTimeout((function(){t.setAttribute("data-animate-fade-out","")}),r),setTimeout((function(){n.removeChild(t)}),r+300))},G=function(e){var t=this;this.currentFileUploads=0,this.isFormLocked=!1,this.reload=function(){var e=t.freeform.form;t.freeform.form.querySelectorAll("[data-freeform-file-upload]").forEach((function(n){n.style.setProperty("--accent",n.dataset.accent),n.addEventListener("dragenter",t.handleDrag(n)),n.addEventListener("dragleave",t.handleDragLeave(n)),n.addEventListener("dragover",t.handleDrag(n)),n.addEventListener("drop",t.handleDrop(n)),n.addEventListener("click",t.handleClick(n)),n.addEventListener(a.onChange,t.handleChanges),function(e,t){if(parseInt(e.dataset.fileCount||"0")){var r=e.querySelector("[data-preview-zone]"),n=e.dataset.freeformFileUpload,o=new FormData(t.form);o.delete("action"),o.append("handle",n);var a=e.getAttribute("data-base-url");A.post("".concat(a,"/files"),o,{headers:{"Freeform-Preflight":!0}}).then((function(o){o.data.forEach((function(o){var i=R(o,t);I(o.extension)&&i.querySelector("[data-thumbnail]").setAttribute("data-has-preview","");var s=new FormData(t.form);s.delete("action"),s.append("handle",n),s.append("id",o.id),i.querySelector("[data-remove-button]").addEventListener("click",(function(){confirm("Are you sure?")&&A.post("".concat(a,"/files/delete"),s).then((function(){r.removeChild(i),D(e)})).catch((function(e){alert(e.message)}))})),i.appendChild(M(n,o)),i.setAttribute("data-completed",""),r.appendChild(i)})),D(e)})).catch(console.error)}}(n,t.freeform),e.addEventListener(r.reset,t.handleReset(n)),n.querySelector("input[type=file]").addEventListener("change",t.handleManualUpload(n))}))},this.handleChanges=function(e){var t=e.container;t.querySelector("[data-preview-zone]").querySelectorAll("[data-file-preview]").length>0?t.setAttribute("data-contains-files",""):t.removeAttribute("data-contains-files")},this.attachDragState=function(e){e instanceof HTMLElement&&(e.dataset.dragging="")},this.detachDragState=function(e){e instanceof HTMLElement&&delete e.dataset.dragging},this.handleDrag=function(e){return function(r){r.preventDefault(),r.stopPropagation(),t.attachDragState(e)}},this.handleDragLeave=function(e){return function(r){r.preventDefault(),r.stopPropagation(),t.detachDragState(e)}},this.handleDrop=function(e){return function(r){r.preventDefault(),r.stopPropagation(),t.detachDragState(e);var n=r.dataTransfer.files;t.initFileUpload(n,e)}},this.handleClick=function(e){return function(){var t=e.querySelector('input[type="file"]');if(!t)throw new Error("File upload corrupted");t.click()}},this.handleManualUpload=function(e){return function(r){var n=r.target,o=n.files;t.initFileUpload(o,e),n.value=null}},this.initFileUpload=function(e,r){for(var n=r.dataset,o=n.freeformFileUpload,a=n.maxFiles,i=n.maxSize,s=r.dataset,l=s.messageSize,c=s.messageFiles,u=r.querySelector("[data-preview-zone]"),d=r.querySelectorAll("[data-file-preview]:not([data-has-errors])").length,f=0;f<e.length;f++){if(d>=parseInt(a)){H(r,c);break}var p=e.item(f);p.size>parseInt(i)?H(r,l):(t.currentFileUploads++,U(p,o,r,u,t.freeform).finally((function(){t.currentFileUploads--,t.handleUploadLockdown()})),d++,t.handleUploadLockdown())}},this.handleReset=function(e){return function(){e.querySelectorAll("[data-file-preview]").forEach((function(e){return e.parentNode.removeChild(e)})),f(a.onChange,{container:e},e)}},this.handleUploadLockdown=function(){t.currentFileUploads>0?t.isFormLocked||(t.isFormLocked=!0,t.freeform.lockSubmit("file-upload")):(t.freeform.unlockSubmit("file-upload"),t.isFormLocked=!1)},this.freeform=e,this.reload()},$=function(e){var t=this;if(this.scriptAdded=!1,this.reload=function(){t.freeform.has("data-scripts-js-mask")&&t.freeform.form.querySelectorAll("*[data-masked-input]").forEach((function(e){var t=e.getAttribute("data-pattern");t&&new IMask(e,{mask:t})}))},this.freeform=e,this.freeform.has("data-scripts-js-mask")&&!this.scriptAdded){var r=document.createElement("script");r.src="https://cdnjs.cloudflare.com/ajax/libs/imask/6.0.7/imask.min.js",r.async=!1,r.defer=!1,r.addEventListener("load",(function(){t.reload()})),document.body.appendChild(r),this.scriptAdded=!0}},W=function(e){var t=this;this.reload=function(){t.freeform.form.querySelectorAll('[data-field-type="rating"]').forEach((function(e){var t=e.dataset,r=t.colorIdle,n=t.colorHover,o=t.colorSelected;e.style.setProperty("--ff-rating-color-idle",r),e.style.setProperty("--ff-rating-color-hover",n),e.style.setProperty("--ff-rating-color-selected",o)}))},this.freeform=e,this.reload()},K=function(e){var t=this;if(this.scriptAdded=!1,this.reload=function(){t.freeform.has("data-scripts-signature")&&t.freeform.form.querySelectorAll("canvas[data-signature-field]").forEach((function(e){var t=e.dataset,r=t.borderColor,n=t.backgroundColor,o=t.penColor,a=t.dotSize;e.style.borderWidth="1px",e.style.borderStyle="solid",e.style.borderColor=r;var i=e.previousSibling,s=e.parentNode.querySelector("[data-signature-clear]"),l=i.value,c=new SignaturePad(e,{onEnd:function(){i.value=c.toDataURL()},backgroundColor:n,penColor:o,dotSize:a,maxWidth:a,throttle:5});if(s&&s.addEventListener("click",(function(){c.clear(),i.value=""})),l){var u=new Image;c.clear(),u.src=l,u.onload=function(){e.getContext("2d").drawImage(u,0,0,e.width,e.height)}}}))},this.freeform=e,this.freeform.has("data-scripts-signature")&&!this.scriptAdded){var r=document.createElement("script");r.src="//cdn.jsdelivr.net/npm/signature_pad@2.3.2/dist/signature_pad.min.js",r.async=!1,r.defer=!1,r.addEventListener("load",(function(){t.reload()})),document.body.appendChild(r),this.scriptAdded=!0}},X=/([^[]+)\[(\d+)\](\[\d+\])$/g,J=/^(labeled-.*)-(\d+)-(\d+)-(\d+)$/g,Y=function(e,t){var r=e.getAttribute("data-max-rows"),n=e.querySelectorAll("tbody > tr").length;r&&t&&(t.style.display=n>=parseInt(r,10)?"none":"block")},Z=function(e){return"string"==typeof e&&(e=e.split(" ")),e},V=function(e,t){Z(t).map((function(t){return e.classList.add(t)}))},Q=function(e,t){Z(t).map((function(t){return e.classList.remove(t)}))};function ee(e){var t;void 0!==e&&(e instanceof HTMLElement&&(null===(t=e.parentElement)||void 0===t||t.removeChild(e)),(e instanceof HTMLCollection||e instanceof NodeList)&&Array.from(e).forEach((function(e){var t;null===(t=e.parentElement)||void 0===t||t.removeChild(e)})))}var te=!1,re=function(e){return function(t){var r=t.target;if(1!==r.closest("tbody").querySelectorAll("tr").length){var n=r.closest("table"),a=r.closest("tr");e._dispatchEvent(o.onRemoveRow,{table:n,row:a}),ee(a),e._dispatchEvent(o.afterRemoveRow,{table:n})}}},ne=function(e){var t=e.form.querySelectorAll("[data-freeform-table]"),r=e.freeform.options.errorClassField;t.forEach((function(e){var t=e.querySelectorAll("thead th[data-column-required]"),n=Array.from(t).map((function(e){return e.cellIndex}));e.querySelectorAll("tbody tr").forEach((function(e){var t=e.querySelectorAll("td");n.forEach((function(e){var n=t[e].querySelector("input, textarea, select"),o=Boolean(n.value);["radio","checkbox"].includes((null==n?void 0:n.type)||"")&&(o=n.checked),n&&o?Q(n,r):"radio"===(null==n?void 0:n.type)?n.parentElement.parentElement.querySelectorAll("input").forEach((function(e){V(e,r)})):V(n,r)}))}))}))},oe=function(e){var t=this;this.reload=function(){var e;(function(e){var t;e.form.querySelectorAll("[data-freeform-table]").forEach((function(r){t=r.parentNode.querySelector("[data-freeform-table-add-row]"),Y(r,t),t&&(e.form.addEventListener(o.afterRemoveRow,(function(){Y(r,t)})),t.addEventListener("click",(function(){var n=r.querySelector("tbody > tr:last-child");if(n){for(var a=n.cloneNode(!0),i=a.querySelectorAll("textarea, input, select"),s=function(){for(var e=r.querySelectorAll("textarea, input, select"),t=0,n=0;n<e.length;n++){var o=e[n],a=X.exec(o.name);if(X.lastIndex=0,a){var i=parseInt(a[2]);t=Math.max(t,i)}}return++t}(),l=0;l<i.length;l++){var c=i[l],u=c.dataset.defaultValue||"";c.name=c.name.replace(X,"$1[".concat(s,"]$3")),c.id&&J.test(c.id)&&(c.id=c.id.replace(J,"$1-".concat(s,"-$3-$4")),c.nextSibling.htmlFor=c.id),"INPUT"===c.tagName&&"radio"===c.type?c.checked=c.value===u:"INPUT"===c.tagName&&"checkbox"===c.type?c.checked="1"===u:c.value=u}e._dispatchEvent(o.onAddRow,{table:r,row:a}),r.querySelector("tbody").appendChild(a),e._dispatchEvent(o.afterRowAdded,{table:r,row:a}),Y(r,t)}})))}))})(t.freeform),(e=t.freeform).form.querySelectorAll("[data-freeform-table]").forEach((function(t){for(var r=t.querySelectorAll("[data-freeform-table-remove-row]"),n=0;n<r.length;n++)r[n].addEventListener("click",re(e))})),te||(te=!0,e.form.addEventListener(o.afterRowAdded,(function(t){var r=t.row.querySelector("[data-freeform-table-remove-row]");r&&r.addEventListener("click",re(e))}))),function(e){e.form.addEventListener(r.renderFieldErrors,ne)}(t.freeform)},this.freeform=e,this.reload()},ae=function(){return ae=Object.assign||function(e){for(var t,r=1,n=arguments.length;r<n;r++)for(var o in t=arguments[r])Object.prototype.hasOwnProperty.call(t,o)&&(e[o]=t[o]);return e},ae.apply(this,arguments)},ie=function(e){if(this.reload=function(){},window.dataLayer=window.dataLayer||[],this.freeform=e,this.form=e.form,this.freeform.has("data-gtm")){var t=this.form.dataset.gtmEventName||"form-submission",n=this.form.dataset.handle;this.form.addEventListener(r.ajaxSuccess,(function(r){var o=r.response,a=e._dispatchEvent("freeform-gtm-data-layer-push",{payload:{},response:o}),i=ae({event:t,form:n,submission:{id:o.submissionId,token:o.submissionToken}},a.payload);window.dataLayer.push(i)}))}},se=function(e,t){if(e===t)return!0;if(null==e||null==t)return e===t;if(Number.isNaN(e)&&Number.isNaN(t))return!0;if(typeof e!=typeof t)return!1;if(Array.isArray(e)&&Array.isArray(t))return e.length===t.length&&e.every((function(e,r){return se(e,t[r])}));if("object"==typeof e&&"object"==typeof t){var r=Object.keys(e),n=Object.keys(t);return r.length===n.length&&r.every((function(r){return t.hasOwnProperty(r)&&se(e[r],t[r])}))}return!1},le=function(e,t,r){if(r||2===arguments.length)for(var n,o=0,a=t.length;o<a;o++)!n&&o in t||(n||(n=Array.prototype.slice.call(t,0,o)),n[o]=t[o]);return e.concat(n||Array.prototype.slice.call(t))},ce=function(e){return function(t){return t.conditions.some((function(t){var r=e.name,n=t.field;return n===r||"".concat(n,"[]")===r}))}},ue=function(e){var t=this;this.reload=function(){var e=t.form.querySelector("[data-rules-json]");if(e){var r=JSON.parse(e.dataset.rulesJson),n=r.rules,o=r.values;t.values=o,0===n.fields.length&&0===n.buttons.length||(Array.from(t.form.elements).forEach((function(e){var r=n.fields.filter(ce(e)),o=n.buttons.filter(ce(e)),a=le(le([],r,!0),o,!0);if(0!==a.length){var i;switch(e.tagName){case"TEXTAREA":case"INPUT":var s=e;if("hidden"===s.type)return;switch(s.type){case"radio":case"checkbox":i="change";break;case"number":i=["keyup","change"];break;default:i="keyup"}break;case"SELECT":i="change"}if(i){var l=Array.isArray(i)?i:[i];a.forEach((function(r){l.forEach((function(o){return e.addEventListener(o,function(e){return function(){t.applyRule(e),n.fields.filter((function(t){return t!==e})).forEach((function(e){return t.applyRule(e)}))}}(r))}))}))}}})),n.fields.forEach((function(e){return t.applyRule(e)})),n.buttons.forEach((function(e){return t.applyRule(e)})))}},this.applyRule=function(e){var r="field"in e?'[data-field-container="'.concat(e.field,'"]'):'[data-button-container="'.concat(e.button,'"]'),o=t.form.querySelector(r);if(o){var a=e.display,i=e.combinator,s=e.conditions,l="and"===i?s.every(t.verifyCondition):s.some(t.verifyCondition);return"show"===a?(o.style.display=l?"":"none",l?delete o.dataset.hidden:o.dataset.hidden=""):(o.style.display=l?"none":"",l?o.dataset.hidden="":delete o.dataset.hidden),f(n.applied,{rule:e},o),!0}},this.verifyCondition=function(e){var r=null,n=document.querySelector('[data-field-container="'.concat(e.field,'"]'));if(n){var o=t.form[e.field]||t.form["".concat(e.field,"[]")],a="checkbox"===n.getAttribute("data-field-type"),i="checkboxes"===n.getAttribute("data-field-type");r=void 0!==n.dataset.hidden?null:a?o[1].checked?"1":"":i?Array.from(o).filter((function(e){return e.checked})).map((function(e){return e.value})):o instanceof HTMLSelectElement&&o.multiple?Array.from(o.options).filter((function(e){return e.selected})).map((function(e){return e.value})):o instanceof RadioNodeList?(r=Array.from(o).filter((function(e){return e.checked})).map((function(e){return e.value}))).length>0?r[0]:"":o.value}else{var s=t.values[e.field]||"";r="boolean"==typeof s?s?"1":"":"number"==typeof s?"".concat(s):s}if("object"==typeof r)switch(e.operator){case"equals":return se(r,[e.value]);case"notEquals":return!se(r,[e.value]);case"contains":return null==r?void 0:r.includes(e.value);case"notContains":return!(null==r?void 0:r.includes(e.value));case"isEmpty":return null===r||0===r.length;case"isNotEmpty":return(null==r?void 0:r.length)>0;case"isOneOf":case"isNotOneOf":var l="isOneOf"===e.operator,c=(e.value?JSON.parse(e.value):[]).map((function(e){return e.toLowerCase()})),u=r.some((function(e){return c.includes(e.toLowerCase())}));return 0===c.length?l?0!==r.length:0===r.length:l?u:!u;default:return!1}switch(e.operator){case"equals":return"".concat(r).toLowerCase()==="".concat(e.value).toLowerCase();case"notEquals":return"".concat(r).toLowerCase()!=="".concat(e.value).toLowerCase();case"greaterThan":return parseFloat(r)>parseFloat(e.value);case"greaterThanOrEquals":return parseFloat(r)>=parseFloat(e.value);case"lessThan":return parseFloat(r)<parseFloat(e.value);case"lessThanOrEquals":return parseFloat(r)<=parseFloat(e.value);case"contains":return"".concat(r).toLowerCase().includes("".concat(e.value).toLowerCase());case"notContains":return!"".concat(r).toLowerCase().includes("".concat(e.value).toLowerCase());case"startsWith":return"".concat(r).toLowerCase().startsWith("".concat(e.value).toLowerCase());case"endsWith":return"".concat(r).toLowerCase().endsWith("".concat(e.value).toLowerCase());case"isEmpty":return null===r||0===r.length;case"isNotEmpty":return(null==r?void 0:r.length)>0;case"isOneOf":case"isNotOneOf":l="isOneOf"===e.operator;var d=(e.value?JSON.parse(e.value):[]).map((function(e){return e.toLowerCase()})),f=d.includes(r.toLowerCase());return 0===d.length?l?0===r.length:0!==r.length:l?f:!f;default:return!1}},this.freeform=e,this.form=e.form,this.reload()},de=function(e){var t=this;this.reload=function(){},this.freeform=e,this.form=e.form,this.form.addEventListener(r.handleActions,(function(e){var r=e.actions.find((function(e){return"save-form"===e.name}));if(r){var n=r.metadata,o=n.key,a=n.token,s=n.url,l=f(i.saveFormhandleToken,{key:o,token:a,url:s},t.form);l.defaultPrevented||(window.location.href=l.url)}}))},fe=function(){return fe=Object.assign||function(e){for(var t,r=1,n=arguments.length;r<n;r++)for(var o in t=arguments[r])Object.prototype.hasOwnProperty.call(t,o)&&(e[o]=t[o]);return e},fe.apply(this,arguments)},pe=function(e,t,r,n){return new(r||(r=Promise))((function(o,a){function i(e){try{l(n.next(e))}catch(e){a(e)}}function s(e){try{l(n.throw(e))}catch(e){a(e)}}function l(e){var t;e.done?o(e.value):(t=e.value,t instanceof r?t:new r((function(e){e(t)}))).then(i,s)}l((n=n.apply(e,t||[])).next())}))},me=function(e,t){var r,n,o,a={label:0,sent:function(){if(1&o[0])throw o[1];return o[1]},trys:[],ops:[]},i=Object.create(("function"==typeof Iterator?Iterator:Object).prototype);return i.next=s(0),i.throw=s(1),i.return=s(2),"function"==typeof Symbol&&(i[Symbol.iterator]=function(){return this}),i;function s(s){return function(l){return function(s){if(r)throw new TypeError("Generator is already executing.");for(;i&&(i=0,s[0]&&(a=0)),a;)try{if(r=1,n&&(o=2&s[0]?n.return:s[0]?n.throw||((o=n.return)&&o.call(n),0):n.next)&&!(o=o.call(n,s[1])).done)return o;switch(n=0,o&&(s=[2&s[0],o.value]),s[0]){case 0:case 1:o=s;break;case 4:return a.label++,{value:s[1],done:!1};case 5:a.label++,n=s[1],s=[0];continue;case 7:s=a.ops.pop(),a.trys.pop();continue;default:if(!((o=(o=a.trys).length>0&&o[o.length-1])||6!==s[0]&&2!==s[0])){a=0;continue}if(3===s[0]&&(!o||s[1]>o[0]&&s[1]<o[3])){a.label=s[1];break}if(6===s[0]&&a.label<o[1]){a.label=o[1],o=s;break}if(o&&a.label<o[2]){a.label=o[2],a.ops.push(s);break}o[2]&&a.ops.pop(),a.trys.pop();continue}s=t.call(e,a)}catch(e){s=[6,e],n=0}finally{r=o=0}if(5&s[0])throw s[1];return{value:s[0]?s[1]:void 0,done:!0}}([s,l])}}},he=function(){function e(n){var o=this;if(this.options={ajax:!1,disableReset:!1,disableSubmit:!1,autoScroll:!1,scrollToAnchor:!1,scrollOffset:0,scrollElement:window,showProcessingSpinner:!1,showProcessingText:!1,processingText:null,prevButtonName:"form_previous_page_button",skipHtmlReload:!1,successBannerMessage:"Form has been submitted successfully!",errorBannerMessage:"Sorry, there was an error submitting the form. Please try again.",errorClassBanner:"freeform-form-errors",errorClassList:"freeform-errors",errorClassField:"freeform-has-errors",successClassBanner:"freeform-form-success",removeMessages:null,renderSuccess:null,renderFormErrors:null,renderFieldErrors:null},this._initializedHandlers=[],this._handlers=[c,ue,d,$,W,K,oe,ie,G,de],this._lockList=new Set,this._disableList=new Set,this._scrollToForm=function(){var e=o.options,t=e.scrollOffset,r=e.scrollElement,n=o.form.getBoundingClientRect().top+window.scrollY+t;r.scrollTo({top:n,behavior:o._isReducedMotion()?"instant":"smooth"})},this._isReducedMotion=function(){var e=window.matchMedia("(prefers-reduced-motion: reduce)");return!e||e.matches},this._setUp=function(){o._attachListeners(),o._getSubmitButtons().forEach((function(e){e.dataset.originalText=e.innerText,e.dataset.processingText=o.options.processingText}))},this._initHandlers=function(){o._handlers.forEach((function(e){o._initializedHandlers.push(new e(o))}))},this._resetHandlers=function(){o._initializedHandlers.forEach((function(e){return e.reload?e.reload():null}))},this.has=function(e){return null!==o.form.getAttribute(e)},this.setOption=function(e,t){o.options[e]=t},this.disableForm=function(){o.form.dataset.freeformDisabled=""},this.enableForm=function(){delete o.form.dataset.freeformDisabled},this.disableSubmit=function(e){void 0===e&&(e="freeform"),o._disableList.add(e);for(var t=0,r=Array.from(o._getSubmitButtons());t<r.length;t++)r[t].disabled=!0},this.enableSubmit=function(e){if(void 0===e&&(e="freeform"),o._disableList.delete(e),!(o._disableList.size>0))for(var t=0,r=Array.from(o._getSubmitButtons());t<r.length;t++)r[t].disabled=!1},this.lockSubmit=function(e){if(void 0===e&&(e="freeform"),o._lockList.add(e),!(o._lockList.size>1)){for(var t=o.options,r=t.disableSubmit,n=t.showProcessingSpinner,a=t.showProcessingText,i=o._getSubmitButtons(),s=0;s<i.length;s++){var l=i[s];r&&(l.disabled=!0)}var c=o._lastButtonPressed;c||(c=o._getSubmitButtons()[0]||void 0),c&&(n&&c.classList.add("freeform-processing"),a&&(c.innerText=c.dataset.processingText))}},this.unlockSubmit=function(e){void 0===e&&(e="freeform"),o._lockList.delete(e),o._lockList.size>0||o._unlockSubmitButtons()},this.forceUnlockSubmit=function(){o._lockList.clear(),o._unlockSubmitButtons()},this.triggerResubmit=function(){o.unlockSubmit(),o._lastButtonPressed?o._lastButtonPressed.click():o.triggerSubmit()},this.triggerSubmit=function(){o.unlockSubmit();var e=o._getSubmitButtons();e.length&&e[0].click()},this._unlockSubmitButtons=function(){for(var e=o.options,t=e.disableSubmit,r=e.showProcessingSpinner,n=e.showProcessingText,a=o._getSubmitButtons(),i=0;i<a.length;i++){var s=a[i];t&&(s.disabled=!1),r&&s.classList.remove("freeform-processing"),n&&(s.innerText=s.dataset.originalText)}},this._setInstances=function(){var t=o.form;e.instances.set(t,o),t.freeform=o},this._attachListeners=function(){var e=o.form,t=o.form.querySelector("input[name=freeform-action]"),n=e.querySelectorAll("[data-freeform-action]");t&&(n.forEach((function(e){return e.addEventListener("click",(function(){o._lastButtonPressed=e,t.value=e.getAttribute("data-freeform-action")}))})),e.addEventListener(r.ajaxAfterSubmit,(function(){t.value="submit"}))),e.addEventListener("submit",o._onSubmit),e.querySelectorAll("input, select, textarea").forEach((function(e){return e.addEventListener("change",(function(e){o._removeMessageFrom(e.target)}))}))},this._onSubmit=function(t){return pe(o,void 0,void 0,(function(){var n,o,a,i,s,l,c;return me(this,(function(u){switch(u.label){case 0:if(this.lockSubmit(),t.preventDefault(),t.stopPropagation(),n=this.options.ajax,o=t.submitter,a=!1,o&&o.name&&o.name===e._BACK_BUTTON_NAME&&(a=!0),i={},this._dispatchEvent(r.submit,{isBackButtonPressed:a,cancelable:!0,addCallback:function(e,t){void 0===t&&(t=0),void 0===i[t]&&(i[t]=[]),i[t].push(e)}}).defaultPrevented)return this.forceUnlockSubmit(),this._dispatchEvent(r.afterFailedSubmit,{cancelable:!1}),[2,!1];s=Object.entries(i).sort((function(e,t){var r=e[0],n=t[0];return Number(r)-Number(n)})).flatMap((function(e){return e[1]})),l=0,c=s,u.label=1;case 1:return l<c.length?[4,(0,c[l])()]:[3,4];case 2:if(!1===u.sent())return this.forceUnlockSubmit(),this._dispatchEvent(r.afterFailedSubmit,{cancelable:!1}),[2,!1];u.label=3;case 3:return l++,[3,1];case 4:return n?(this._onSubmitAjax(t),[2,!1]):(this.form.submit(),[2])}}))}))},this._removeMessages=function(){if(!o._dispatchEvent(r.removeMessages).defaultPrevented){if("function"==typeof o.options.removeMessages)return o.options.removeMessages=o.options.removeMessages.bind(o),o.options.removeMessages();var e=o,t=e.form,n=e.options,a=n.successClassBanner,i=n.errorClassBanner,s=n.errorClassList,l=n.errorClassField;ee(t.querySelectorAll(".".concat(Z(s).join(".")))),t.querySelectorAll(".".concat(Z(l).join("."))).forEach((function(e){o._removeMessageFrom(e)})),ee(t.querySelectorAll(B(a))),ee(document.querySelectorAll(B(i)))}},this._removeMessageFrom=function(e){if(!o._dispatchEvent(r.fieldRemoveMessages,{field:e}).defaultPrevented){var t=o.options,n=t.errorClassList,a=t.errorClassField,i=e.parentNode;e.type&&("radio"===e.type||"checkbox"===e.type&&/\[]$/.test(e.name))&&(i=e.parentNode.parentNode),ee(i.querySelector(B(n)));for(var s=i.querySelectorAll("input, select, textarea"),l=0;l<s.length;l++)Q(s[l],a)}},this._renderSuccessBanner=function(){if(!o._dispatchEvent(r.renderSuccess).defaultPrevented){if("function"==typeof o.options.renderSuccess)return o.options.renderSuccess=o.options.renderSuccess.bind(o),o.options.renderSuccess();var e=o,t=e.form,n=e.options,a=n.successBannerMessage,i=n.successClassBanner,s=document.createElement("div");V(s,i);var l=document.createElement("p");l.appendChild(document.createTextNode(a)),s.appendChild(l),t.insertBefore(s,t.childNodes[0])}},this._renderFieldErrors=function(e){if(o._dispatchEvent(r.renderFieldErrors,{errors:e}).defaultPrevented)return!1;if("function"==typeof o.options.renderFieldErrors)return o.options.renderFieldErrors=o.options.renderFieldErrors.bind(o),o.options.renderFieldErrors(e);var t=o,n=t.form,a=t.options,i=a.errorClassList,s=a.errorClassField;for(var l in e){var c=e[l],u=document.createElement("ul");u.setAttribute("data-field-errors",""),V(u,i);for(var d=0;d<c.length;d++){var f=c[d],p=document.createElement("li");p.appendChild(document.createTextNode(f)),u.appendChild(p)}var m=n.querySelector('[data-field-container="'.concat(l,'"]')),h=n.querySelector('[data-error-append-target="'.concat(l,'"]')),v=n.querySelectorAll('\n          [name="'.concat(l,'"],\n          [type=file][name="').concat(l,'"],\n          [type=file][name="').concat(l,'[]"],\n          [type=checkbox][name="').concat(l,'[]"],\n          [type=radio][name="').concat(l,'"],\n          select[multiple][name="').concat(l,'[]"],\n          [data-freeform-file-upload="').concat(l,'"]\n        '));if(!m)return;for(var b=0;b<v.length;b++){var y=v[b];V(y,s)}h?h.appendChild(u):m.appendChild(u)}},this._renderFormErrors=function(e){if(o._dispatchEvent(r.renderFormErrors,{errors:e}).defaultPrevented)return!1;if("function"==typeof o.options.renderFormErrors)return o.options.renderFormErrors=o.options.renderFormErrors.bind(o),o.options.renderFormErrors(e);var t=o,n=t.form,a=t.options,i=a.errorClassBanner,s=a.errorBannerMessage,l=document.createElement("div");V(l,i);var c=document.createElement("p");if(c.appendChild(document.createTextNode(s)),l.appendChild(c),e.length){for(var u=document.createElement("ul"),d=0;d<e.length;d++){var f=e[d],p=document.createElement("li");p.appendChild(document.createTextNode(f)),u.appendChild(p)}l.appendChild(u)}n.insertBefore(l,n.childNodes[0])},this._prepareFormData=function(){var e=o.form,t=new FormData(e);if(navigator.userAgent.indexOf("Safari")>-1)for(var r=0;r<e.elements.length;r++){var n=e.elements[r];"file"===n.type&&""===n.value&&t.delete(n.name)}return t},this.quickSave=function(e,t){return pe(o,void 0,void 0,(function(){var n,o,a,i,s,l,c,u,d,f;return me(this,(function(p){switch(p.label){case 0:n=this.form,(o=this._prepareFormData()).set("action","freeform/submit/quick-save"),o.set("storage-secret",e),t&&o.set("token",t),p.label=1;case 1:return p.trys.push([1,3,,4]),[4,A(n.getAttribute("action")||window.location.href,{method:n.getAttribute("method"),data:o})];case 2:return a=p.sent(),[3,4];case 3:return i=p.sent(),417===(null===(f=null==i?void 0:i.response)||void 0===f?void 0:f.status)?(this.unlockSubmit(),[2,!1]):[3,4];case 4:if(this._removeMessages(),s=a.data,200===a.status){if(l=s.success,c=s.errors,u=s.formErrors,d=s.storageToken,l)return[2,d];(c||u)&&(this._dispatchEvent(r.ajaxError,{request:a,response:s,errors:c,formErrors:u}),this._dispatchEvent(r.afterFailedSubmit,{cancelable:!1}),this._renderFieldErrors(c),this._renderFormErrors(u)),this.options.autoScroll&&this._scrollToForm()}else this._dispatchEvent(r.ajaxError,{request:a,response:s}),this._dispatchEvent(r.afterFailedSubmit,{cancelable:!1});return this.unlockSubmit(),[2]}}))}))},this._onSubmitAjax=function(e){var n=o.form,a=o._prepareFormData(),i=new XMLHttpRequest,s=e.submitter;s&&s.name&&a.append(s.name,"1");var l=n.getAttribute("method")||"POST",c=n.getAttribute("action")||window.location.href;o._dispatchEvent(r.ajaxBeforeSubmit,{data:a,request:i}).defaultPrevented||A(c,{data:a,method:l,request:i}).then((function(e){if(o._removeMessages(),200===e.status){var s=(v=e.data).success,l=v.finished,c=v.actions,u=void 0===c?[]:c,d=v.errors,f=v.formErrors,p=v.returnUrl;if(o._dispatchEvent(r.ajaxBeforeSuccess,{request:i,response:v}).defaultPrevented)return;if(u.length)o._dispatchEvent(r.handleActions,{response:v,actions:u,cancelable:!1});else if(s){if(l&&v.onSuccess===t.RedirectReturnUrl&&p){if(o._dispatchEvent(r.ajaxSuccess,{request:i,response:v}).defaultPrevented)return;return void(window.location.href=p)}null===v.html||o.options.skipHtmlReload||(n.innerHTML=v.html.replace(/<form[^>]*>/,"").replace("</form>","")),o.options.skipHtmlReload||(o._resetHandlers(),o._setUp()),l&&(o.options.disableReset||(n.reset(),o._dispatchEvent(r.reset)),v.onSuccess===t.Reload&&o._renderSuccessBanner()),o._dispatchEvent(r.ajaxSuccess,{request:i,response:v})}else(d||f)&&(o._dispatchEvent(r.ajaxError,{request:i,response:v,errors:d,formErrors:f}),o._dispatchEvent(r.afterFailedSubmit,{cancelable:!1}),o._renderFieldErrors(d),o._renderFormErrors(f));var m=null==v?void 0:v.freeform_payload;if(m){var h=n.querySelector("input[name^=freeform_payload]");h&&(h.value=m)}o._dispatchEvent(r.ajaxAfterSubmit,{data:a,request:i,response:v,cancelable:!1}),o.options.autoScroll&&o._scrollToForm()}else{var v=i.response;o._dispatchEvent(r.ajaxError,{request:i,response:v})}o.unlockSubmit()}))},this._getSubmitButtons=function(){var e=o.form.querySelectorAll("*[type=submit][data-freeform-action]");return e.length?e:o.form.querySelectorAll("*[type=submit]")},this._getBackButtons=function(){return o.form.querySelectorAll('*[type=submit][data-freeform-action="back"]')},this._dispatchEvent=function(e,t,r){var n=f(e,fe(fe({},t),{form:o.form,freeform:o}),r);return document.dispatchEvent(n),o.form.dispatchEvent(n),n},e.instances.get(n))return e.instances.get(n);this.id=n.dataset.id,this.form=n,this._setInstances();var a={ajax:null!==n.getAttribute("data-ajax"),disableReset:null!==n.getAttribute("data-disable-reset"),scrollToAnchor:null!==n.getAttribute("data-scroll-to-anchor"),autoScroll:null!==n.getAttribute("data-auto-scroll"),disableSubmit:null!==n.getAttribute("data-disable-submit"),showProcessingSpinner:null!==n.getAttribute("data-show-processing-spinner"),showProcessingText:null!==n.getAttribute("data-show-processing-text"),processingText:n.getAttribute("data-processing-text"),successBannerMessage:n.getAttribute("data-success-message"),errorBannerMessage:n.getAttribute("data-error-message"),skipHtmlReload:null!==n.getAttribute("data-skip-html-reload")};this.options=fe(fe({},this.options),a),this.disableSubmit("init");var i=setInterval((function(){return pe(o,void 0,void 0,(function(){var e;return me(this,(function(t){return"complete"===document.readyState&&(clearInterval(i),e=this._dispatchEvent(r.ready,{options:{}}),this.options=fe(fe({},this.options),e.options),this._setUp(),this._initHandlers(),this.enableSubmit("init"),this.options.scrollToAnchor&&this._scrollToForm()),[2]}))}))}),50)}return e._BACK_BUTTON_NAME="form_previous_page_button",e.instances=new WeakMap,e.getInstance=function(t){return e.instances.get(t)},e}(),ve=he;document.querySelectorAll("form[data-freeform]").forEach((function(e){new he(e)}));var be,ye=function(e){var t;"FORM"===e.nodeName&&void 0!==(null===(t=e.dataset)||void 0===t?void 0:t.freeform)&&new he(e),null==e||e.childNodes.forEach(ye)},ge=new MutationObserver((function(e){e.forEach((function(e){"childList"===e.type&&e.addedNodes.forEach((function(e){ye(e)}))}))})),we=0,Ee=function(){if(we>25)return console.warn("Freeform observer timed out"),clearTimeout(be);document.body?ge.observe(document.body,{childList:!0,subtree:!0}):(we++,be&&clearTimeout(be),be=setTimeout(Ee,50))};Ee()}();
