+++ packages/scripts/src/lib/plugin/helpers/ajax/ajax.classes.ts
@@ -1,0 +1,30 @@
+export class CancelToken {
+  private cancelFn: () => void;
+
+  constructor() {}
+
+  cancel() {
+    if (this.cancelFn) {
+      this.cancelFn();
+      this.cancelFn = null;
+    }
+  }
+
+  _setCancelFn(fn: () => void) {
+    this.cancelFn = fn;
+  }
+}
+
+export class HttpError<T> extends Error {
+  response: XMLHttpRequest & { data: T };
+  status: number;
+
+  constructor(message: string, response: XMLHttpRequest, data: T) {
+    super(message);
+    this.response = { ...response, data };
+    this.status = response.status;
+
+    // Set the prototype explicitly to maintain the prototype chain
+    Object.setPrototypeOf(this, HttpError.prototype);
+  }
+}
